Clarify spec boundaries by moving helper and test noise out of root modules
The spec runtime and execution entry files had become the last place where
support logic, adapter stubs, bridge helpers, and inline tests still crowded
the orchestration paths. This rewrite consolidates the remaining spec cleanup
into one reviewable decision so future readers see a clean boundary story
instead of checkpoint noise.

Constraint: Base branch already absorbed the earlier app and daemon cleanup slices from the original PR branch
Rejected: Preserve the stale mixed-scope branch shape verbatim | would reintroduce unrelated drift and larger review surface
Confidence: high
Scope-risk: moderate
Reversibility: clean
Directive: Keep crates/spec/src/spec_runtime.rs and crates/spec/src/spec_execution.rs orchestration-focused; grow helpers and tests under crates/spec/src/spec_runtime/* and crates/spec/src/spec_execution/* first
Tested: ./scripts/cargo-local-toolchain.sh fmt --all -- --check; ./scripts/check_architecture_boundaries.sh; ./scripts/cargo-local-toolchain.sh test --workspace (follow-up rerun in progress on refreshed branch)
Not-tested: Full all-features and clippy gates are still pending on the refreshed branch
