+++ crates/spec/src/spec_execution.rs
@@ -1,24 +1,21 @@
 use std::{
-    collections::{BTreeMap, BTreeSet},
+    collections::BTreeMap,
     ffi::OsString,
-    fs,
-    path::{Path, PathBuf},
     sync::{Arc, Mutex},
     time::{SystemTime, UNIX_EPOCH},
 };
 
 use kernel::{
     ArchitectureBoundaryPolicy, ArchitectureGuardReport, AuditEventKind, AutoProvisionAgent,
-    AutoProvisionRequest, BootstrapPolicy, BootstrapReport, BootstrapTaskStatus,
-    BridgeSupportMatrix, Clock, CodebaseAwarenessConfig, CodebaseAwarenessEngine,
-    CodebaseAwarenessSnapshot, ConnectorCommand, InMemoryAuditSink, IntegrationCatalog,
-    LoongKernel, MemoryCoreRequest, MemoryExtensionRequest, PluginAbsorbReport,
-    PluginActivationPlan, PluginActivationStatus, PluginBootstrapExecutor, PluginBridgeKind,
-    PluginCompatibility, PluginCompatibilityShimSupport, PluginDescriptor, PluginScanReport,
-    PluginScanner, PluginSetup, PluginSetupReadinessContext, PluginSlotClaim,
-    PluginTranslationReport, PluginTranslator, ProvisionPlan, RuntimeCoreRequest,
-    RuntimeExtensionRequest, StaticPolicyEngine, SystemClock, TaskIntent, TaskSupervisor,
-    ToolCoreRequest, ToolExtensionRequest, plugin_bridge_is_high_risk_auto_apply,
+    AutoProvisionRequest, BootstrapPolicy, BootstrapReport, BootstrapTaskStatus, Clock,
+    CodebaseAwarenessConfig, CodebaseAwarenessEngine, CodebaseAwarenessSnapshot, ConnectorCommand,
+    InMemoryAuditSink, IntegrationCatalog, LoongKernel, MemoryCoreRequest, MemoryExtensionRequest,
+    PluginAbsorbReport, PluginActivationPlan, PluginActivationStatus, PluginBootstrapExecutor,
+    PluginCompatibility, PluginCompatibilityShimSupport, PluginScanReport, PluginScanner,
+    PluginSetup, PluginSetupReadinessContext, PluginSlotClaim, PluginTranslationReport,
+    PluginTranslator, ProvisionPlan, RuntimeCoreRequest, RuntimeExtensionRequest,
+    StaticPolicyEngine, SystemClock, TaskIntent, TaskSupervisor, ToolCoreRequest,
+    ToolExtensionRequest, plugin_bridge_is_high_risk_auto_apply,
     plugin_provenance_summary_for_descriptor,
 };
 use serde_json::{Value, json};
@@ -29,22 +26,57 @@ use crate::programmatic::execute_programmatic_tool_call;
 use crate::spec_runtime::*;
 
 mod approval_policy;
+mod bootstrap_support;
+mod bridge_runtime_policy_support;
+mod bridge_selection_support;
 mod bridge_support_policy;
+mod integration_catalog_support;
+mod path_policy_support;
 mod plugin_inventory;
+mod plugin_kind_support;
+mod plugin_metadata;
 mod plugin_preflight;
 mod plugin_preflight_policy;
+mod reporting;
 mod security_scan_eval;
 mod security_scan_policy;
 mod tool_search;
 use approval_policy::evaluate_approval_guard;
+pub(crate) use bootstrap_support::current_epoch_s;
+use bootstrap_support::{apply_default_selection, bootstrap_policy};
+#[cfg(test)]
+pub(crate) use bridge_runtime_policy_support::collect_verified_env_var_names;
+use bridge_runtime_policy_support::{
+    bridge_runtime_policy, filter_scan_report_by_activation, filter_scan_report_by_keys,
+    resolve_plugin_setup_readiness_context,
+};
+use bridge_selection_support::bridge_support_matrix;
 use bridge_support_policy::bridge_support_policy_checksum;
+pub(crate) use integration_catalog_support::fnv1a64_hex;
+pub use integration_catalog_support::hex_lower;
+use integration_catalog_support::{
+    default_integration_catalog, operation_connector_name, register_dynamic_catalog_connectors,
+    snapshot_runtime_integration_catalog,
+};
+pub(crate) use path_policy_support::{
+    normalize_allowed_path_prefixes, validate_wasm_guest_readable_config_keys,
+};
+pub use path_policy_support::{normalize_path_for_policy, resolve_plugin_relative_path};
 use plugin_inventory::execute_plugin_inventory;
+pub(crate) use plugin_kind_support::descriptor_bridge_kind;
+use plugin_metadata::enrich_scan_report_with_translation;
 use plugin_preflight::execute_plugin_preflight;
+use reporting::{
+    build_blocked_spec_run_report, build_plugin_trust_summary, emit_plugin_trust_audit_event,
+    emit_security_scan_audit_event,
+};
 use security_scan_eval::evaluate_plugin_security_scan;
 use security_scan_policy::{
     apply_security_scan_delta, emit_security_scan_siem_record, security_scan_process_allowlist,
 };
-use tool_search::execute_tool_search;
+use tool_search::{
+    build_tool_search_operation_summary, emit_tool_search_audit_event, execute_tool_search,
+};
 
 pub use approval_policy::operation_risk_profile;
 pub use bridge_support_policy::{
@@ -832,318 +864,6 @@ async fn execute_spec_internal(
     }
 }
 
-fn blocked_outcome(reason: &str) -> Value {
-    json!({
-        "status": "blocked",
-        "reason": reason,
-    })
-}
-
-#[allow(clippy::too_many_arguments)]
-fn build_blocked_spec_run_report(
-    pack_id: String,
-    agent_id: String,
-    reason: String,
-    approval_guard: ApprovalDecisionReport,
-    bridge_support_source: Option<String>,
-    bridge_support_checksum: Option<String>,
-    bridge_support_sha256: Option<String>,
-    bridge_support_delta_source: Option<String>,
-    bridge_support_delta_sha256: Option<String>,
-    self_awareness: Option<CodebaseAwarenessSnapshot>,
-    architecture_guard: Option<ArchitectureGuardReport>,
-    plugin_scan_reports: Vec<PluginScanReport>,
-    plugin_translation_reports: Vec<PluginTranslationReport>,
-    plugin_activation_plans: Vec<PluginActivationPlan>,
-    plugin_bootstrap_reports: Vec<BootstrapReport>,
-    plugin_bootstrap_queue: Vec<String>,
-    plugin_absorb_reports: Vec<PluginAbsorbReport>,
-    security_scan_report: Option<SecurityScanReport>,
-    auto_provision_plan: Option<ProvisionPlan>,
-    integration_catalog: IntegrationCatalog,
-    include_audit: bool,
-    audit_sink: &Arc<InMemoryAuditSink>,
-) -> SpecRunReport {
-    let plugin_trust_summary = build_plugin_trust_summary(
-        &plugin_scan_reports,
-        &plugin_activation_plans,
-        &plugin_bootstrap_reports,
-    );
-    SpecRunReport {
-        schema_version: SPEC_RUN_REPORT_SCHEMA_VERSION,
-        schema: json_schema_descriptor(
-            SPEC_RUN_REPORT_SCHEMA_VERSION,
-            SPEC_RUN_REPORT_SCHEMA_SURFACE,
-            SPEC_RUN_REPORT_SCHEMA_PURPOSE,
-        ),
-        pack_id,
-        agent_id,
-        operation_kind: "blocked",
-        blocked_reason: Some(reason.clone()),
-        approval_guard,
-        bridge_support_source,
-        bridge_support_checksum,
-        bridge_support_sha256,
-        bridge_support_delta_source,
-        bridge_support_delta_sha256,
-        self_awareness,
-        architecture_guard,
-        plugin_scan_reports,
-        plugin_translation_reports,
-        plugin_activation_plans,
-        plugin_bootstrap_reports,
-        plugin_trust_summary,
-        tool_search_summary: None,
-        plugin_bootstrap_queue,
-        plugin_absorb_reports,
-        security_scan_report,
-        auto_provision_plan,
-        outcome: blocked_outcome(&reason),
-        integration_catalog,
-        audit_events: if include_audit {
-            Some(audit_sink.snapshot())
-        } else {
-            None
-        },
-    }
-}
-
-fn build_plugin_trust_summary(
-    plugin_scan_reports: &[PluginScanReport],
-    plugin_activation_plans: &[PluginActivationPlan],
-    plugin_bootstrap_reports: &[BootstrapReport],
-) -> PluginTrustSummary {
-    let mut summary = PluginTrustSummary::default();
-    let mut provenance_by_plugin = BTreeMap::new();
-    let mut bootstrap_by_plugin = BTreeMap::new();
-
-    for report in plugin_scan_reports {
-        for descriptor in &report.descriptors {
-            summary.scanned_plugins = summary.scanned_plugins.saturating_add(1);
-            match descriptor.manifest.trust_tier {
-                kernel::PluginTrustTier::Official => {
-                    summary.official_plugins = summary.official_plugins.saturating_add(1);
-                }
-                kernel::PluginTrustTier::VerifiedCommunity => {
-                    summary.verified_community_plugins =
-                        summary.verified_community_plugins.saturating_add(1);
-                }
-                kernel::PluginTrustTier::Unverified => {
-                    summary.unverified_plugins = summary.unverified_plugins.saturating_add(1);
-                }
-            }
-            provenance_by_plugin.insert(
-                (
-                    descriptor.path.clone(),
-                    descriptor.manifest.plugin_id.clone(),
-                ),
-                plugin_provenance_summary_for_descriptor(descriptor),
-            );
-        }
-    }
-
-    for report in plugin_bootstrap_reports {
-        for task in &report.tasks {
-            bootstrap_by_plugin.insert(
-                (task.source_path.clone(), task.plugin_id.clone()),
-                (task.status, task.reason.clone()),
-            );
-        }
-    }
-
-    for plan in plugin_activation_plans {
-        for candidate in &plan.candidates {
-            if plugin_bridge_is_high_risk_auto_apply(candidate.bridge_kind) {
-                summary.high_risk_plugins = summary.high_risk_plugins.saturating_add(1);
-            }
-            if !matches!(candidate.trust_tier, kernel::PluginTrustTier::Unverified)
-                || !plugin_bridge_is_high_risk_auto_apply(candidate.bridge_kind)
-            {
-                continue;
-            }
-
-            summary.high_risk_unverified_plugins =
-                summary.high_risk_unverified_plugins.saturating_add(1);
-
-            let plugin_key = (candidate.source_path.clone(), candidate.plugin_id.clone());
-            let provenance_summary = provenance_by_plugin
-                .get(&plugin_key)
-                .cloned()
-                .unwrap_or_else(|| {
-                    kernel::format_plugin_provenance_summary(
-                        candidate.source_kind,
-                        &candidate.source_path,
-                        candidate.package_manifest_path.as_deref(),
-                    )
-                });
-            let (bootstrap_status, bootstrap_reason) = bootstrap_by_plugin
-                .get(&plugin_key)
-                .map(|(status, reason)| (Some(*status), Some(reason.clone())))
-                .unwrap_or((None, None));
-
-            if matches!(
-                bootstrap_status,
-                Some(BootstrapTaskStatus::DeferredUnsupportedAutoApply)
-            ) && bootstrap_reason
-                .as_deref()
-                .is_some_and(|reason| reason.contains("bootstrap trust policy"))
-            {
-                summary.blocked_auto_apply_plugins =
-                    summary.blocked_auto_apply_plugins.saturating_add(1);
-            }
-
-            summary
-                .review_required_plugins
-                .push(PluginTrustReviewEntry {
-                    plugin_id: candidate.plugin_id.clone(),
-                    source_path: candidate.source_path.clone(),
-                    provenance_summary,
-                    trust_tier: candidate.trust_tier,
-                    bridge_kind: candidate.bridge_kind,
-                    activation_status: candidate.status,
-                    bootstrap_status,
-                    reason: bootstrap_reason.unwrap_or_else(|| candidate.reason.clone()),
-                });
-        }
-    }
-
-    summary.review_required_plugins.sort_by(|left, right| {
-        left.plugin_id
-            .cmp(&right.plugin_id)
-            .then_with(|| left.source_path.cmp(&right.source_path))
-    });
-
-    summary
-}
-
-fn build_tool_search_operation_summary(outcome: &Value) -> Option<ToolSearchOperationSummary> {
-    let payload = outcome.as_object()?;
-    let results = payload.get("results")?.as_array()?;
-    let top_results = results
-        .iter()
-        .take(3)
-        .filter_map(build_tool_search_operation_summary_entry)
-        .collect::<Vec<_>>();
-    let trust_filter_summary = payload
-        .get("trust_filter_summary")
-        .cloned()
-        .and_then(|value| serde_json::from_value::<ToolSearchTrustFilterSummary>(value).ok())
-        .unwrap_or_default();
-    let query = payload
-        .get("query")
-        .and_then(Value::as_str)
-        .unwrap_or_default()
-        .to_owned();
-    let returned = payload
-        .get("returned")
-        .and_then(Value::as_u64)
-        .map_or(results.len(), |value| value as usize);
-    let trust_tiers = payload
-        .get("trust_tiers")
-        .and_then(Value::as_array)
-        .map(|tiers| {
-            tiers
-                .iter()
-                .filter_map(Value::as_str)
-                .map(str::to_owned)
-                .collect::<Vec<_>>()
-        })
-        .unwrap_or_default();
-
-    Some(ToolSearchOperationSummary {
-        headline: build_tool_search_operation_headline(
-            &query,
-            returned,
-            &trust_tiers,
-            &trust_filter_summary,
-            &top_results,
-        ),
-        query,
-        returned,
-        trust_tiers,
-        trust_filter_summary,
-        top_results,
-    })
-}
-
-fn build_tool_search_operation_summary_entry(
-    value: &Value,
-) -> Option<ToolSearchOperationSummaryEntry> {
-    let entry = value.as_object()?;
-    Some(ToolSearchOperationSummaryEntry {
-        tool_id: entry.get("tool_id")?.as_str()?.to_owned(),
-        provider_id: entry.get("provider_id")?.as_str()?.to_owned(),
-        connector_name: entry.get("connector_name")?.as_str()?.to_owned(),
-        trust_tier: entry
-            .get("trust_tier")
-            .and_then(Value::as_str)
-            .map(str::to_owned),
-        bridge_kind: entry.get("bridge_kind")?.as_str()?.to_owned(),
-        score: entry
-            .get("score")
-            .and_then(Value::as_u64)
-            .map_or(0, |value| value as u32),
-        setup_ready: entry
-            .get("setup_ready")
-            .and_then(Value::as_bool)
-            .unwrap_or(false),
-        deferred: entry
-            .get("deferred")
-            .and_then(Value::as_bool)
-            .unwrap_or(false),
-        loaded: entry
-            .get("loaded")
-            .and_then(Value::as_bool)
-            .unwrap_or(false),
-    })
-}
-
-fn build_tool_search_operation_headline(
-    query: &str,
-    returned: usize,
-    trust_tiers: &[String],
-    trust_filter_summary: &ToolSearchTrustFilterSummary,
-    top_results: &[ToolSearchOperationSummaryEntry],
-) -> String {
-    let result_noun = if returned == 1 { "result" } else { "results" };
-    let mut parts = vec![format!("returned {returned} {result_noun}")];
-
-    if trust_filter_summary.applied {
-        let scope = if trust_filter_summary.effective_tiers.is_empty() {
-            "none".to_owned()
-        } else {
-            trust_filter_summary.effective_tiers.join(",")
-        };
-        parts.push(format!("trust_scope={scope}"));
-        if trust_filter_summary.filtered_out_candidates > 0 {
-            let filtered_noun = if trust_filter_summary.filtered_out_candidates == 1 {
-                "candidate"
-            } else {
-                "candidates"
-            };
-            parts.push(format!(
-                "filtered_out={} {filtered_noun}",
-                trust_filter_summary.filtered_out_candidates
-            ));
-        }
-        if trust_filter_summary.conflicting_requested_tiers {
-            parts.push("conflicting_trust_filters=true".to_owned());
-        }
-    } else if !trust_tiers.is_empty() {
-        parts.push(format!("requested_tiers={}", trust_tiers.join(",")));
-    }
-
-    if let Some(first) = top_results.first() {
-        parts.push(format!("top_match={}", first.provider_id));
-    }
-
-    if query.is_empty() {
-        parts.join("; ")
-    } else {
-        format!("query=\"{query}\"; {}", parts.join("; "))
-    }
-}
-
 #[derive(Debug, Default)]
 struct SecurityScanDelta {
     scanned_plugins: usize,
@@ -1152,1124 +872,6 @@ struct SecurityScanDelta {
     low_findings: usize,
     findings: Vec<SecurityFinding>,
     block_reason: Option<String>,
-}
-
-fn emit_security_scan_audit_event(
-    kernel: &LoongKernel<StaticPolicyEngine>,
-    pack_id: &str,
-    agent_id: &str,
-    report: &SecurityScanReport,
-) -> Result<(), String> {
-    if report.scanned_plugins == 0 && report.total_findings == 0 {
-        return Ok(());
-    }
-
-    let categories: Vec<String> = report
-        .findings
-        .iter()
-        .map(|finding| finding.category.clone())
-        .collect::<BTreeSet<_>>()
-        .into_iter()
-        .collect();
-    let finding_ids: Vec<String> = report
-        .findings
-        .iter()
-        .map(|finding| finding.correlation_id.clone())
-        .collect::<BTreeSet<_>>()
-        .into_iter()
-        .collect();
-
-    kernel
-        .record_audit_event(
-            Some(agent_id),
-            AuditEventKind::SecurityScanEvaluated {
-                pack_id: pack_id.to_owned(),
-                scanned_plugins: report.scanned_plugins,
-                total_findings: report.total_findings,
-                high_findings: report.high_findings,
-                medium_findings: report.medium_findings,
-                low_findings: report.low_findings,
-                blocked: report.blocked,
-                block_reason: report.block_reason.clone(),
-                categories,
-                finding_ids,
-            },
-        )
-        .map_err(|error| format!("failed to record security scan audit event: {error}"))
-}
-
-fn emit_plugin_trust_audit_event(
-    kernel: &LoongKernel<StaticPolicyEngine>,
-    pack_id: &str,
-    agent_id: &str,
-    summary: &PluginTrustSummary,
-) -> Result<(), String> {
-    if summary.scanned_plugins == 0 {
-        return Ok(());
-    }
-
-    let review_required_plugin_ids: Vec<String> = summary
-        .review_required_plugins
-        .iter()
-        .map(|entry| entry.plugin_id.clone())
-        .collect::<BTreeSet<_>>()
-        .into_iter()
-        .collect();
-    let review_required_bridges: Vec<String> = summary
-        .review_required_plugins
-        .iter()
-        .map(|entry| entry.bridge_kind.as_str().to_owned())
-        .collect::<BTreeSet<_>>()
-        .into_iter()
-        .collect();
-
-    kernel
-        .record_audit_event(
-            Some(agent_id),
-            AuditEventKind::PluginTrustEvaluated {
-                pack_id: pack_id.to_owned(),
-                scanned_plugins: summary.scanned_plugins,
-                official_plugins: summary.official_plugins,
-                verified_community_plugins: summary.verified_community_plugins,
-                unverified_plugins: summary.unverified_plugins,
-                high_risk_plugins: summary.high_risk_plugins,
-                high_risk_unverified_plugins: summary.high_risk_unverified_plugins,
-                blocked_auto_apply_plugins: summary.blocked_auto_apply_plugins,
-                review_required_plugin_ids,
-                review_required_bridges,
-            },
-        )
-        .map_err(|error| format!("failed to record plugin trust audit event: {error}"))
-}
-
-fn emit_tool_search_audit_event(
-    kernel: &LoongKernel<StaticPolicyEngine>,
-    pack_id: &str,
-    agent_id: &str,
-    summary: &ToolSearchOperationSummary,
-) -> Result<(), String> {
-    let top_provider_ids = summary
-        .top_results
-        .iter()
-        .map(|entry| entry.provider_id.clone())
-        .collect::<Vec<_>>();
-
-    kernel
-        .record_audit_event(
-            Some(agent_id),
-            AuditEventKind::ToolSearchEvaluated {
-                pack_id: pack_id.to_owned(),
-                query: summary.query.clone(),
-                returned: summary.returned,
-                trust_filter_applied: summary.trust_filter_summary.applied,
-                query_requested_tiers: summary.trust_filter_summary.query_requested_tiers.clone(),
-                structured_requested_tiers: summary
-                    .trust_filter_summary
-                    .structured_requested_tiers
-                    .clone(),
-                effective_tiers: summary.trust_filter_summary.effective_tiers.clone(),
-                conflicting_requested_tiers: summary
-                    .trust_filter_summary
-                    .conflicting_requested_tiers,
-                filtered_out_candidates: summary.trust_filter_summary.filtered_out_candidates,
-                filtered_out_tier_counts: summary
-                    .trust_filter_summary
-                    .filtered_out_tier_counts
-                    .clone(),
-                top_provider_ids,
-            },
-        )
-        .map_err(|error| format!("failed to record tool search audit event: {error}"))
-}
-
-pub fn resolve_plugin_relative_path(source_path: &str, artifact: &str) -> PathBuf {
-    let candidate = PathBuf::from(artifact);
-    if candidate.is_absolute() {
-        return candidate;
-    }
-
-    let source = Path::new(source_path);
-    if let Some(parent) = source.parent() {
-        parent.join(candidate)
-    } else {
-        candidate
-    }
-}
-
-fn normalize_allowed_path_prefixes(prefixes: &[String]) -> Vec<PathBuf> {
-    prefixes
-        .iter()
-        .map(|prefix| normalize_path_for_policy(&PathBuf::from(prefix)))
-        .collect()
-}
-
-fn validate_wasm_guest_readable_config_keys(keys: &BTreeSet<String>) -> CliResult<()> {
-    for key in keys {
-        let key_is_supported = wasm_guest_config_key_is_supported(key.as_str());
-        if key_is_supported {
-            continue;
-        }
-
-        return Err(format!(
-            "invalid security scan runtime.guest_readable_config_keys entry `{key}`: entries must use `{WASM_GUEST_CONFIG_PROVIDER_PREFIX}<metadata_key>` or `{WASM_GUEST_CONFIG_CHANNEL_PREFIX}<metadata_key>`"
-        ));
-    }
-
-    Ok(())
-}
-
-pub fn normalize_path_for_policy(path: &Path) -> PathBuf {
-    if let Ok(canonical) = fs::canonicalize(path) {
-        return canonical;
-    }
-
-    if path.is_absolute() {
-        return path.to_path_buf();
-    }
-
-    std::env::current_dir()
-        .map(|cwd| cwd.join(path))
-        .unwrap_or_else(|_| path.to_path_buf())
-}
-
-fn descriptor_bridge_kind(descriptor: &PluginDescriptor) -> PluginBridgeKind {
-    if let Some(raw) = descriptor.manifest.metadata.get("bridge_kind")
-        && let Some(kind) = parse_bridge_kind_label(raw)
-    {
-        return kind;
-    }
-
-    let language = descriptor.language.trim().to_ascii_lowercase();
-    match language.as_str() {
-        "wasm" | "wat" => return PluginBridgeKind::WasmComponent,
-        "rust" | "go" | "c" | "cpp" | "cxx" => return PluginBridgeKind::NativeFfi,
-        "python" | "javascript" | "typescript" | "java" => return PluginBridgeKind::ProcessStdio,
-        _ => {}
-    }
-
-    if let Some(endpoint) = descriptor.manifest.endpoint.as_deref() {
-        let endpoint_lower = endpoint.to_ascii_lowercase();
-        if endpoint_lower.starts_with("http://") || endpoint_lower.starts_with("https://") {
-            return PluginBridgeKind::HttpJson;
-        }
-        if endpoint_lower.ends_with(".wasm") {
-            return PluginBridgeKind::WasmComponent;
-        }
-    }
-
-    PluginBridgeKind::Unknown
-}
-
-fn bridge_support_matrix(spec: &RunnerSpec) -> (BridgeSupportMatrix, bool) {
-    match &spec.bridge_support {
-        Some(bridge) if bridge.enabled => {
-            (bridge_support_spec_matrix(bridge), bridge.enforce_supported)
-        }
-        _ => (BridgeSupportMatrix::default(), false),
-    }
-}
-
-fn resolve_plugin_setup_readiness_context<I>(
-    readiness_spec: Option<&PluginSetupReadinessSpec>,
-    env_vars: I,
-) -> PluginSetupReadinessContext
-where
-    I: IntoIterator<Item = (OsString, OsString)>,
-{
-    let Some(readiness_spec) = readiness_spec else {
-        let verified_env_vars = collect_verified_env_var_names(env_vars);
-
-        return PluginSetupReadinessContext {
-            verified_env_vars,
-            verified_config_keys: BTreeSet::new(),
-        };
-    };
-
-    let mut verified_env_vars = BTreeSet::new();
-    if readiness_spec.inherit_process_env {
-        verified_env_vars = collect_verified_env_var_names(env_vars);
-    }
-
-    let explicit_verified_env_vars = collect_verified_name_list(&readiness_spec.verified_env_vars);
-    verified_env_vars.extend(explicit_verified_env_vars);
-
-    let verified_config_keys = collect_verified_name_list(&readiness_spec.verified_config_keys);
-
-    PluginSetupReadinessContext {
-        verified_env_vars,
-        verified_config_keys,
-    }
-}
-
-fn collect_verified_env_var_names<I>(env_vars: I) -> BTreeSet<String>
-where
-    I: IntoIterator<Item = (OsString, OsString)>,
-{
-    let mut verified_env_vars = BTreeSet::new();
-
-    for (raw_name, raw_value) in env_vars {
-        let name = raw_name.to_string_lossy().into_owned();
-        let trimmed_name = name.trim();
-        let name_is_blank = trimmed_name.is_empty();
-        if name_is_blank {
-            continue;
-        }
-
-        let value = raw_value.to_string_lossy().into_owned();
-        let trimmed_value = value.trim();
-        let value_is_blank = trimmed_value.is_empty();
-        if value_is_blank {
-            continue;
-        }
-
-        verified_env_vars.insert(name);
-    }
-
-    verified_env_vars
-}
-
-fn collect_verified_name_list(values: &[String]) -> BTreeSet<String> {
-    let mut verified_names = BTreeSet::new();
-
-    for raw_value in values {
-        let value = raw_value.trim().to_owned();
-        if value.is_empty() {
-            continue;
-        }
-
-        verified_names.insert(value);
-    }
-
-    verified_names
-}
-
-pub(super) fn bridge_support_spec_matrix(bridge: &BridgeSupportSpec) -> BridgeSupportMatrix {
-    let mut matrix = BridgeSupportMatrix::default();
-    if !bridge.supported_bridges.is_empty() {
-        matrix.supported_bridges = bridge.supported_bridges.iter().copied().collect();
-    }
-    if !bridge.supported_adapter_families.is_empty() {
-        matrix.supported_adapter_families =
-            bridge.supported_adapter_families.iter().cloned().collect();
-    }
-    if !bridge.supported_compatibility_modes.is_empty() {
-        matrix.supported_compatibility_modes = bridge
-            .supported_compatibility_modes
-            .iter()
-            .copied()
-            .collect();
-    }
-    if !bridge.supported_compatibility_shims.is_empty() {
-        matrix.supported_compatibility_shims = bridge
-            .supported_compatibility_shims
-            .iter()
-            .cloned()
-            .collect();
-    }
-    if !bridge.supported_compatibility_shim_profiles.is_empty() {
-        matrix.supported_compatibility_shim_profiles = bridge
-            .supported_compatibility_shim_profiles
-            .iter()
-            .cloned()
-            .map(PluginCompatibilityShimSupport::normalized)
-            .map(|profile| (profile.shim.clone(), profile))
-            .collect();
-        matrix
-            .supported_compatibility_shims
-            .extend(matrix.supported_compatibility_shim_profiles.keys().cloned());
-    }
-    matrix
-}
-
-fn raw_bridge_runtime_spec(spec: &RunnerSpec) -> SecurityRuntimeExecutionSpec {
-    let raw_runtime = spec
-        .bridge_support
-        .as_ref()
-        .filter(|bridge| bridge.enabled)
-        .and_then(|bridge| bridge.security_scan.as_ref())
-        .map(|scan| scan.runtime.clone());
-
-    raw_runtime.unwrap_or_default()
-}
-
-fn bridge_runtime_policy(
-    spec: &RunnerSpec,
-    security_scan: Option<&SecurityScanSpec>,
-) -> CliResult<BridgeRuntimePolicy> {
-    let Some(bridge) = &spec.bridge_support else {
-        return Ok(BridgeRuntimePolicy::default());
-    };
-    if !bridge.enabled {
-        return Ok(BridgeRuntimePolicy::default());
-    }
-
-    let runtime = security_scan
-        .map(|scan| scan.runtime.clone())
-        .unwrap_or_default();
-    let raw_runtime = raw_bridge_runtime_spec(spec);
-    let bridge_circuit_breaker = if security_scan.is_some() {
-        runtime.bridge_circuit_breaker.clone()
-    } else {
-        raw_runtime.bridge_circuit_breaker
-    };
-    validate_connector_circuit_breaker_policy(
-        &bridge_circuit_breaker,
-        "bridge runtime circuit breaker",
-    )?;
-    let (compatibility_matrix, _) = bridge_support_matrix(spec);
-    let (wasm_require_hash_pin, wasm_required_sha256_by_plugin) = security_scan
-        .map(|scan| {
-            (
-                scan.wasm.require_hash_pin,
-                scan.wasm.required_sha256_by_plugin.clone(),
-            )
-        })
-        .unwrap_or_else(|| (false, BTreeMap::new()));
-    let wasm_allowed_path_prefixes = runtime
-        .allowed_path_prefixes
-        .iter()
-        .map(PathBuf::from)
-        .map(|path| normalize_path_for_policy(&path))
-        .collect();
-    let wasm_guest_readable_config_keys =
-        collect_verified_name_list(&runtime.guest_readable_config_keys);
-    validate_wasm_guest_readable_config_keys(&wasm_guest_readable_config_keys)?;
-
-    Ok(BridgeRuntimePolicy {
-        execute_process_stdio: bridge.execute_process_stdio,
-        execute_http_json: bridge.execute_http_json,
-        execute_wasm_component: runtime.execute_wasm_component,
-        compatibility_matrix,
-        allowed_process_commands: bridge
-            .allowed_process_commands
-            .iter()
-            .map(|value| value.trim().to_ascii_lowercase())
-            .filter(|value| !value.is_empty())
-            .collect(),
-        bridge_circuit_breaker,
-        wasm_allowed_path_prefixes,
-        wasm_guest_readable_config_keys,
-        wasm_max_component_bytes: runtime.max_component_bytes,
-        wasm_max_output_bytes: runtime.max_output_bytes,
-        wasm_fuel_limit: runtime.fuel_limit,
-        wasm_timeout_ms: runtime.timeout_ms,
-        wasm_require_hash_pin,
-        wasm_required_sha256_by_plugin,
-        enforce_execution_success: bridge.enforce_execution_success,
-    })
-}
-
-pub fn current_epoch_s() -> u64 {
-    SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .map(|duration| duration.as_secs())
-        .unwrap_or(0)
-}
-
-fn bootstrap_policy(spec: &RunnerSpec) -> Option<BootstrapPolicy> {
-    let bootstrap = spec.bootstrap.as_ref()?;
-    if !bootstrap.enabled {
-        return None;
-    }
-
-    let mut policy = BootstrapPolicy::default();
-    if let Some(value) = bootstrap.allow_http_json_auto_apply {
-        policy.allow_http_json_auto_apply = value;
-    }
-    if let Some(value) = bootstrap.allow_process_stdio_auto_apply {
-        policy.allow_process_stdio_auto_apply = value;
-    }
-    if let Some(value) = bootstrap.allow_native_ffi_auto_apply {
-        policy.allow_native_ffi_auto_apply = value;
-    }
-    if let Some(value) = bootstrap.allow_wasm_component_auto_apply {
-        policy.allow_wasm_component_auto_apply = value;
-    }
-    if let Some(value) = bootstrap.allow_mcp_server_auto_apply {
-        policy.allow_mcp_server_auto_apply = value;
-    }
-    if let Some(value) = bootstrap.allow_acp_bridge_auto_apply {
-        policy.allow_acp_bridge_auto_apply = value;
-    }
-    if let Some(value) = bootstrap.allow_acp_runtime_auto_apply {
-        policy.allow_acp_runtime_auto_apply = value;
-    }
-    if let Some(value) = bootstrap.block_unverified_high_risk_auto_apply {
-        policy.block_unverified_high_risk_auto_apply = value;
-    }
-    if let Some(value) = bootstrap.enforce_ready_execution {
-        policy.enforce_ready_execution = value;
-    }
-    if let Some(value) = bootstrap.max_tasks {
-        policy.max_tasks = value.max(1);
-    }
-    Some(policy)
-}
-
-fn filter_scan_report_by_activation(
-    report: &PluginScanReport,
-    activation: &PluginActivationPlan,
-) -> PluginScanReport {
-    let ready_keys: BTreeSet<(String, String)> = activation
-        .candidates
-        .iter()
-        .filter(|candidate| matches!(candidate.status, PluginActivationStatus::Ready))
-        .map(|candidate| (candidate.source_path.clone(), candidate.plugin_id.clone()))
-        .collect();
-
-    filter_scan_report_by_keys(report, &ready_keys)
-}
-
-fn filter_scan_report_by_keys(
-    report: &PluginScanReport,
-    allowed_keys: &BTreeSet<(String, String)>,
-) -> PluginScanReport {
-    let descriptors: Vec<PluginDescriptor> = report
-        .descriptors
-        .iter()
-        .filter(|descriptor| {
-            allowed_keys.contains(&(
-                descriptor.path.clone(),
-                descriptor.manifest.plugin_id.clone(),
-            ))
-        })
-        .cloned()
-        .collect();
-    let diagnostic_findings = report
-        .diagnostic_findings
-        .iter()
-        .filter(|finding| {
-            let (Some(source_path), Some(plugin_id)) =
-                (finding.source_path.as_deref(), finding.plugin_id.as_deref())
-            else {
-                return false;
-            };
-
-            allowed_keys.contains(&(source_path.to_owned(), plugin_id.to_owned()))
-        })
-        .cloned()
-        .collect();
-
-    PluginScanReport {
-        scanned_files: report.scanned_files,
-        matched_plugins: descriptors.len(),
-        diagnostic_findings,
-        descriptors,
-    }
-}
-
-#[derive(Clone)]
-struct PluginTranslationMetadataSnapshot {
-    bridge_kind: String,
-    adapter_family: String,
-    entrypoint_hint: String,
-    source_language: String,
-    channel_id: Option<String>,
-    channel_bridge_transport_family: Option<String>,
-    channel_bridge_target_contract: Option<String>,
-    channel_bridge_account_scope: Option<String>,
-    channel_bridge_ready: Option<bool>,
-    channel_bridge_missing_fields: Vec<String>,
-}
-
-fn enrich_scan_report_with_translation(
-    report: &PluginScanReport,
-    translation: &PluginTranslationReport,
-    activation: Option<&PluginActivationPlan>,
-) -> PluginScanReport {
-    let mut runtime_by_key: BTreeMap<(String, String), PluginTranslationMetadataSnapshot> =
-        BTreeMap::new();
-    let mut activation_contracts_by_key: BTreeMap<
-        (String, String),
-        PluginActivationRuntimeContract,
-    > = BTreeMap::new();
-
-    for entry in &translation.entries {
-        let channel_bridge = entry.channel_bridge.as_ref();
-        let channel_id = channel_bridge
-            .and_then(|bridge| bridge.channel_id.clone())
-            .or_else(|| entry.channel_id.clone());
-        let channel_bridge_transport_family =
-            channel_bridge.and_then(|bridge| bridge.transport_family.clone());
-        let channel_bridge_target_contract =
-            channel_bridge.and_then(|bridge| bridge.target_contract.clone());
-        let channel_bridge_account_scope =
-            channel_bridge.and_then(|bridge| bridge.account_scope.clone());
-        let channel_bridge_ready = channel_bridge.map(|bridge| bridge.readiness.ready);
-        let channel_bridge_missing_fields = channel_bridge
-            .map(|bridge| bridge.readiness.missing_fields.clone())
-            .unwrap_or_default();
-
-        runtime_by_key.insert(
-            (entry.source_path.clone(), entry.plugin_id.clone()),
-            PluginTranslationMetadataSnapshot {
-                bridge_kind: entry.runtime.bridge_kind.as_str().to_owned(),
-                adapter_family: entry.runtime.adapter_family.clone(),
-                entrypoint_hint: entry.runtime.entrypoint_hint.clone(),
-                source_language: entry.runtime.source_language.clone(),
-                channel_id,
-                channel_bridge_transport_family,
-                channel_bridge_target_contract,
-                channel_bridge_account_scope,
-                channel_bridge_ready,
-                channel_bridge_missing_fields,
-            },
-        );
-    }
-
-    if let Some(activation) = activation {
-        for entry in &translation.entries {
-            let Some(candidate) = activation.candidate_for(&entry.source_path, &entry.plugin_id)
-            else {
-                continue;
-            };
-            if !matches!(candidate.status, PluginActivationStatus::Ready) {
-                continue;
-            }
-
-            activation_contracts_by_key.insert(
-                (entry.source_path.clone(), entry.plugin_id.clone()),
-                PluginActivationRuntimeContract {
-                    plugin_id: entry.plugin_id.clone(),
-                    source_path: entry.source_path.clone(),
-                    source_kind: entry.source_kind,
-                    dialect: entry.dialect,
-                    dialect_version: entry.dialect_version.clone(),
-                    compatibility_mode: entry.compatibility_mode,
-                    compatibility_shim: candidate.compatibility_shim.clone(),
-                    bridge_kind: entry.runtime.bridge_kind,
-                    adapter_family: entry.runtime.adapter_family.clone(),
-                    entrypoint_hint: entry.runtime.entrypoint_hint.clone(),
-                    source_language: entry.runtime.source_language.clone(),
-                    compatibility: entry.compatibility.clone(),
-                },
-            );
-        }
-    }
-
-    let descriptors: Vec<PluginDescriptor> = report
-        .descriptors
-        .iter()
-        .cloned()
-        .map(|mut descriptor| {
-            stamp_plugin_provenance_metadata(&mut descriptor);
-            descriptor.manifest.metadata.insert(
-                "plugin_id".to_owned(),
-                descriptor.manifest.plugin_id.clone(),
-            );
-            descriptor.manifest.metadata.insert(
-                "defer_loading".to_owned(),
-                descriptor.manifest.defer_loading.to_string(),
-            );
-            let setup = descriptor.manifest.setup.clone();
-            insert_plugin_setup_metadata(&mut descriptor.manifest.metadata, setup.as_ref());
-            insert_plugin_slot_claims_metadata(
-                &mut descriptor.manifest.metadata,
-                &descriptor.manifest.slot_claims,
-            );
-            let manifest_api_version = descriptor.manifest.api_version.clone();
-            let plugin_version = descriptor.manifest.version.clone();
-            insert_plugin_manifest_contract_metadata(
-                &mut descriptor.manifest.metadata,
-                manifest_api_version,
-                plugin_version,
-            );
-            insert_plugin_compatibility_metadata(
-                &mut descriptor.manifest.metadata,
-                descriptor.manifest.compatibility.as_ref(),
-            );
-            if let Some(summary) = descriptor.manifest.summary.clone() {
-                descriptor
-                    .manifest
-                    .metadata
-                    .entry("summary".to_owned())
-                    .or_insert(summary);
-            }
-            if !descriptor.manifest.tags.is_empty()
-                && let Ok(tags_json) = serde_json::to_string(&descriptor.manifest.tags)
-            {
-                descriptor
-                    .manifest
-                    .metadata
-                    .entry("tags_json".to_owned())
-                    .or_insert(tags_json);
-            }
-            if !descriptor.manifest.input_examples.is_empty()
-                && let Ok(input_examples_json) =
-                    serde_json::to_string(&descriptor.manifest.input_examples)
-            {
-                descriptor
-                    .manifest
-                    .metadata
-                    .entry("input_examples_json".to_owned())
-                    .or_insert(input_examples_json);
-            }
-            if !descriptor.manifest.output_examples.is_empty()
-                && let Ok(output_examples_json) =
-                    serde_json::to_string(&descriptor.manifest.output_examples)
-            {
-                descriptor
-                    .manifest
-                    .metadata
-                    .entry("output_examples_json".to_owned())
-                    .or_insert(output_examples_json);
-            }
-            if let Some(component) = descriptor.manifest.metadata.get("component").cloned() {
-                let resolved = resolve_plugin_relative_path(&descriptor.path, &component);
-                let normalized = normalize_path_for_policy(&resolved);
-                descriptor.manifest.metadata.insert(
-                    "component_resolved_path".to_owned(),
-                    normalized.display().to_string(),
-                );
-            } else {
-                descriptor
-                    .manifest
-                    .metadata
-                    .remove("component_resolved_path");
-            }
-
-            if let Some(runtime_snapshot) = runtime_by_key.get(&(
-                descriptor.path.clone(),
-                descriptor.manifest.plugin_id.clone(),
-            )) {
-                let bridge_kind = runtime_snapshot.bridge_kind.clone();
-                let adapter_family = runtime_snapshot.adapter_family.clone();
-                let entrypoint_hint = runtime_snapshot.entrypoint_hint.clone();
-                let source_language = runtime_snapshot.source_language.clone();
-                descriptor
-                    .manifest
-                    .metadata
-                    .entry("bridge_kind".to_owned())
-                    .or_insert(bridge_kind);
-                descriptor
-                    .manifest
-                    .metadata
-                    .entry("adapter_family".to_owned())
-                    .or_insert(adapter_family);
-                descriptor
-                    .manifest
-                    .metadata
-                    .entry("entrypoint_hint".to_owned())
-                    .or_insert(entrypoint_hint);
-                descriptor
-                    .manifest
-                    .metadata
-                    .entry("source_language".to_owned())
-                    .or_insert(source_language);
-
-                insert_plugin_channel_bridge_metadata(
-                    &mut descriptor.manifest.metadata,
-                    Some(runtime_snapshot),
-                );
-            } else {
-                insert_plugin_channel_bridge_metadata(&mut descriptor.manifest.metadata, None);
-            }
-            insert_plugin_activation_runtime_contract_metadata(
-                &mut descriptor.manifest.metadata,
-                activation_contracts_by_key.get(&(
-                    descriptor.path.clone(),
-                    descriptor.manifest.plugin_id.clone(),
-                )),
-            );
-            descriptor
-        })
-        .collect();
-
-    PluginScanReport {
-        scanned_files: report.scanned_files,
-        matched_plugins: descriptors.len(),
-        diagnostic_findings: report.diagnostic_findings.clone(),
-        descriptors,
-    }
-}
-
-fn insert_plugin_activation_runtime_contract_metadata(
-    metadata: &mut BTreeMap<String, String>,
-    contract: Option<&PluginActivationRuntimeContract>,
-) {
-    let Some(contract) = contract else {
-        metadata.remove(PLUGIN_ACTIVATION_RUNTIME_CONTRACT_METADATA_KEY);
-        metadata.remove(PLUGIN_ACTIVATION_RUNTIME_CONTRACT_CHECKSUM_METADATA_KEY);
-        return;
-    };
-
-    let Ok(serialized) = plugin_activation_runtime_contract_json(contract) else {
-        metadata.remove(PLUGIN_ACTIVATION_RUNTIME_CONTRACT_METADATA_KEY);
-        metadata.remove(PLUGIN_ACTIVATION_RUNTIME_CONTRACT_CHECKSUM_METADATA_KEY);
-        return;
-    };
-
-    metadata.insert(
-        PLUGIN_ACTIVATION_RUNTIME_CONTRACT_METADATA_KEY.to_owned(),
-        serialized.clone(),
-    );
-    metadata.insert(
-        PLUGIN_ACTIVATION_RUNTIME_CONTRACT_CHECKSUM_METADATA_KEY.to_owned(),
-        activation_runtime_contract_checksum_hex(serialized.as_bytes()),
-    );
-}
-
-fn stamp_plugin_provenance_metadata(descriptor: &mut PluginDescriptor) {
-    let source_path_key = "plugin_source_path".to_owned();
-    let source_path_value = descriptor.path.clone();
-    let source_kind_key = "plugin_source_kind".to_owned();
-    let source_kind_value = descriptor.source_kind.as_str().to_owned();
-    let dialect_key = "plugin_dialect".to_owned();
-    let dialect_value = descriptor.dialect.as_str().to_owned();
-    let compatibility_mode_key = "plugin_compatibility_mode".to_owned();
-    let compatibility_mode_value = descriptor.compatibility_mode.as_str().to_owned();
-    let package_root_key = "plugin_package_root".to_owned();
-    let package_root_value = descriptor.package_root.clone();
-    let package_manifest_path_value = descriptor.package_manifest_path.clone();
-    let provenance_summary_key = "plugin_provenance_summary".to_owned();
-    let provenance_summary_value = plugin_provenance_summary_for_descriptor(descriptor);
-    let trust_tier_key = "plugin_trust_tier".to_owned();
-    let trust_tier_value = descriptor.manifest.trust_tier.as_str().to_owned();
-    let metadata = &mut descriptor.manifest.metadata;
-
-    metadata.insert(source_path_key, source_path_value);
-    metadata.insert(source_kind_key, source_kind_value);
-    metadata.insert(dialect_key, dialect_value);
-    metadata.insert(compatibility_mode_key, compatibility_mode_value);
-    metadata.insert(package_root_key, package_root_value);
-    metadata.insert(provenance_summary_key, provenance_summary_value);
-    metadata.insert(trust_tier_key, trust_tier_value);
-
-    if let Some(shim) = kernel::PluginCompatibilityShim::for_mode(descriptor.compatibility_mode) {
-        metadata.insert("plugin_compatibility_shim_id".to_owned(), shim.shim_id);
-        metadata.insert("plugin_compatibility_shim_family".to_owned(), shim.family);
-    } else {
-        metadata.remove("plugin_compatibility_shim_id");
-        metadata.remove("plugin_compatibility_shim_family");
-    }
-
-    if let Some(dialect_version) = descriptor.dialect_version.clone() {
-        metadata.insert("plugin_dialect_version".to_owned(), dialect_version);
-    } else {
-        metadata.remove("plugin_dialect_version");
-    }
-
-    if let Some(package_manifest_path_value) = package_manifest_path_value {
-        let package_manifest_path_key = "plugin_package_manifest_path".to_owned();
-
-        metadata.insert(package_manifest_path_key, package_manifest_path_value);
-    } else {
-        metadata.remove("plugin_package_manifest_path");
-    }
-}
-
-fn insert_plugin_setup_metadata(
-    metadata: &mut BTreeMap<String, String>,
-    setup: Option<&PluginSetup>,
-) {
-    let Some(setup) = setup else {
-        metadata.remove("plugin_setup_mode");
-        metadata.remove("plugin_setup_surface");
-        metadata.remove("plugin_setup_required_env_vars_json");
-        metadata.remove("plugin_setup_recommended_env_vars_json");
-        metadata.remove("plugin_setup_required_config_keys_json");
-        metadata.remove("plugin_setup_default_env_var");
-        metadata.remove("plugin_setup_docs_urls_json");
-        metadata.remove("plugin_setup_remediation");
-        return;
-    };
-
-    let mode_key = "plugin_setup_mode".to_owned();
-    let mode_value = setup.mode.as_str().to_owned();
-    metadata.insert(mode_key, mode_value);
-
-    if let Some(surface) = setup.surface.clone() {
-        let surface_key = "plugin_setup_surface".to_owned();
-        metadata.insert(surface_key, surface);
-    } else {
-        metadata.remove("plugin_setup_surface");
-    }
-
-    insert_plugin_setup_string_list_metadata(
-        metadata,
-        "plugin_setup_required_env_vars_json",
-        &setup.required_env_vars,
-    );
-    insert_plugin_setup_string_list_metadata(
-        metadata,
-        "plugin_setup_recommended_env_vars_json",
-        &setup.recommended_env_vars,
-    );
-    insert_plugin_setup_string_list_metadata(
-        metadata,
-        "plugin_setup_required_config_keys_json",
-        &setup.required_config_keys,
-    );
-
-    if let Some(default_env_var) = setup.default_env_var.clone() {
-        let default_env_var_key = "plugin_setup_default_env_var".to_owned();
-        metadata.insert(default_env_var_key, default_env_var);
-    } else {
-        metadata.remove("plugin_setup_default_env_var");
-    }
-
-    insert_plugin_setup_string_list_metadata(
-        metadata,
-        "plugin_setup_docs_urls_json",
-        &setup.docs_urls,
-    );
-
-    if let Some(remediation) = setup.remediation.clone() {
-        let remediation_key = "plugin_setup_remediation".to_owned();
-        metadata.insert(remediation_key, remediation);
-    } else {
-        metadata.remove("plugin_setup_remediation");
-    }
-}
-
-fn insert_plugin_channel_bridge_metadata(
-    metadata: &mut BTreeMap<String, String>,
-    snapshot: Option<&PluginTranslationMetadataSnapshot>,
-) {
-    let Some(snapshot) = snapshot else {
-        remove_plugin_channel_bridge_metadata(metadata);
-        return;
-    };
-
-    upsert_or_remove_metadata_value(metadata, "plugin_channel_id", snapshot.channel_id.as_ref());
-    upsert_or_remove_metadata_value(
-        metadata,
-        "plugin_channel_bridge_transport_family",
-        snapshot.channel_bridge_transport_family.as_ref(),
-    );
-    upsert_or_remove_metadata_value(
-        metadata,
-        "plugin_channel_bridge_target_contract",
-        snapshot.channel_bridge_target_contract.as_ref(),
-    );
-    upsert_or_remove_metadata_value(
-        metadata,
-        "plugin_channel_bridge_account_scope",
-        snapshot.channel_bridge_account_scope.as_ref(),
-    );
-
-    if let Some(channel_bridge_ready) = snapshot.channel_bridge_ready {
-        let ready_key = "plugin_channel_bridge_ready".to_owned();
-        let ready_value = channel_bridge_ready.to_string();
-        metadata.insert(ready_key, ready_value);
-    } else {
-        metadata.remove("plugin_channel_bridge_ready");
-    }
-
-    upsert_or_remove_json_string_list_metadata(
-        metadata,
-        "plugin_channel_bridge_missing_fields_json",
-        &snapshot.channel_bridge_missing_fields,
-    );
-}
-
-fn remove_plugin_channel_bridge_metadata(metadata: &mut BTreeMap<String, String>) {
-    metadata.remove("plugin_channel_id");
-    metadata.remove("plugin_channel_bridge_transport_family");
-    metadata.remove("plugin_channel_bridge_target_contract");
-    metadata.remove("plugin_channel_bridge_account_scope");
-    metadata.remove("plugin_channel_bridge_ready");
-    metadata.remove("plugin_channel_bridge_missing_fields_json");
-}
-
-fn upsert_or_remove_metadata_value(
-    metadata: &mut BTreeMap<String, String>,
-    key: &str,
-    value: Option<&String>,
-) {
-    let Some(value) = value else {
-        metadata.remove(key);
-        return;
-    };
-
-    let metadata_key = key.to_owned();
-    let metadata_value = value.clone();
-    metadata.insert(metadata_key, metadata_value);
-}
-
-fn upsert_or_remove_json_string_list_metadata(
-    metadata: &mut BTreeMap<String, String>,
-    key: &str,
-    values: &[String],
-) {
-    let serialized = serde_json::to_string(values);
-    let Ok(serialized) = serialized else {
-        metadata.remove(key);
-        return;
-    };
-
-    let metadata_key = key.to_owned();
-    metadata.insert(metadata_key, serialized);
-}
-
-fn insert_plugin_setup_string_list_metadata(
-    metadata: &mut BTreeMap<String, String>,
-    key: &str,
-    values: &[String],
-) {
-    let is_empty = values.is_empty();
-
-    if is_empty {
-        metadata.remove(key);
-        return;
-    }
-
-    let serialized = serde_json::to_string(values);
-    let Ok(serialized) = serialized else {
-        metadata.remove(key);
-        return;
-    };
-
-    let metadata_key = key.to_owned();
-    metadata.insert(metadata_key, serialized);
-}
-
-fn insert_plugin_slot_claims_metadata(
-    metadata: &mut BTreeMap<String, String>,
-    slot_claims: &[PluginSlotClaim],
-) {
-    if slot_claims.is_empty() {
-        metadata.remove("plugin_slot_claims_json");
-        return;
-    }
-
-    if let Ok(serialized) = serde_json::to_string(slot_claims) {
-        metadata.insert("plugin_slot_claims_json".to_owned(), serialized);
-    }
-}
-
-fn insert_plugin_manifest_contract_metadata(
-    metadata: &mut BTreeMap<String, String>,
-    manifest_api_version: Option<String>,
-    plugin_version: Option<String>,
-) {
-    if let Some(api_version) = manifest_api_version {
-        metadata.insert("plugin_manifest_api_version".to_owned(), api_version);
-    } else {
-        metadata.remove("plugin_manifest_api_version");
-    }
-
-    if let Some(version) = plugin_version {
-        metadata.insert("plugin_version".to_owned(), version);
-    } else {
-        metadata.remove("plugin_version");
-    }
-}
-
-fn insert_plugin_compatibility_metadata(
-    metadata: &mut BTreeMap<String, String>,
-    compatibility: Option<&PluginCompatibility>,
-) {
-    let Some(compatibility) = compatibility else {
-        metadata.remove("plugin_compatibility_host_api");
-        metadata.remove("plugin_compatibility_host_version_req");
-        return;
-    };
-
-    if let Some(host_api) = compatibility.host_api.clone() {
-        metadata.insert("plugin_compatibility_host_api".to_owned(), host_api);
-    } else {
-        metadata.remove("plugin_compatibility_host_api");
-    }
-
-    if let Some(host_version_req) = compatibility.host_version_req.clone() {
-        metadata.insert(
-            "plugin_compatibility_host_version_req".to_owned(),
-            host_version_req,
-        );
-    } else {
-        metadata.remove("plugin_compatibility_host_version_req");
-    }
-}
-
-fn fnv1a64_hex(bytes: &[u8]) -> String {
-    activation_runtime_contract_checksum_hex(bytes)
-}
-
-pub fn hex_lower(bytes: &[u8]) -> String {
-    let mut encoded = String::with_capacity(bytes.len().saturating_mul(2));
-    for byte in bytes {
-        encoded.push_str(&format!("{byte:02x}"));
-    }
-    encoded
-}
-
-fn default_integration_catalog() -> IntegrationCatalog {
-    let mut catalog = IntegrationCatalog::new();
-    for (provider_id, connector, version, class) in [
-        ("openai", "openai", "1.0.0", "llm"),
-        ("anthropic", "anthropic", "1.0.0", "llm"),
-        ("github", "github", "1.0.0", "devops"),
-        ("slack", "slack", "1.0.0", "messaging"),
-        ("notion", "notion", "1.0.0", "workspace"),
-    ] {
-        catalog.register_template(kernel::ProviderTemplate {
-            provider_id: provider_id.to_owned(),
-            default_connector_name: connector.to_owned(),
-            default_version: version.to_owned(),
-            metadata: BTreeMap::from([("class".to_owned(), class.to_owned())]),
-        });
-    }
-    catalog
-}
-
-fn register_dynamic_catalog_connectors(
-    kernel: &mut LoongKernel<StaticPolicyEngine>,
-    catalog: Arc<Mutex<IntegrationCatalog>>,
-    bridge_runtime_policy: BridgeRuntimePolicy,
-) {
-    let snapshot = {
-        let guard = match catalog.lock() {
-            Ok(g) => g,
-            Err(_) => return,
-        };
-        guard.providers()
-    };
-
-    for provider in snapshot {
-        let connector = DynamicCatalogConnector::new(
-            provider.connector_name,
-            provider.provider_id,
-            catalog.clone(),
-            bridge_runtime_policy.clone(),
-        );
-
-        kernel.register_core_connector_adapter(connector);
-    }
-}
-
-fn snapshot_runtime_integration_catalog(
-    catalog: &Arc<Mutex<IntegrationCatalog>>,
-) -> Result<IntegrationCatalog, String> {
-    let guard = catalog
-        .lock()
-        .map_err(|_err| "integration catalog mutex poisoned".to_owned())?;
-    let snapshot = guard.clone();
-
-    Ok(snapshot)
-}
-
-fn operation_connector_name(operation: &OperationSpec) -> Option<String> {
-    #[allow(clippy::wildcard_enum_match_arm)]
-    match operation {
-        OperationSpec::ConnectorLegacy { connector_name, .. }
-        | OperationSpec::ConnectorCore { connector_name, .. }
-        | OperationSpec::ConnectorExtension { connector_name, .. } => Some(connector_name.clone()),
-        OperationSpec::ProgrammaticToolCall { steps, .. } => {
-            steps.iter().find_map(|step| match step {
-                ProgrammaticStep::ConnectorCall { connector_name, .. } => {
-                    Some(connector_name.clone())
-                }
-                ProgrammaticStep::ConnectorBatch { calls, .. } => {
-                    calls.first().map(|call| call.connector_name.clone())
-                }
-                ProgrammaticStep::SetLiteral { .. }
-                | ProgrammaticStep::JsonPointer { .. }
-                | ProgrammaticStep::Conditional { .. } => None,
-            })
-        }
-        _ => None,
-    }
 }
 
 async fn execute_spec_operation(
@@ -2689,886 +1291,18 @@ async fn execute_spec_operation(
     }
 }
 
-fn apply_default_selection(
-    kernel: &mut LoongKernel<StaticPolicyEngine>,
-    defaults: Option<&DefaultCoreSelection>,
-) -> CliResult<()> {
-    if let Some(defaults) = defaults {
-        if let Some(connector) = defaults.connector.as_deref() {
-            kernel
-                .set_default_core_connector_adapter(connector)
-                .map_err(|error| {
-                    format!("invalid default connector core adapter ({connector}): {error}")
-                })?;
-        }
-        if let Some(runtime) = defaults.runtime.as_deref() {
-            kernel
-                .set_default_core_runtime_adapter(runtime)
-                .map_err(|error| {
-                    format!("invalid default runtime core adapter ({runtime}): {error}")
-                })?;
-        }
-        if let Some(tool) = defaults.tool.as_deref() {
-            kernel
-                .set_default_core_tool_adapter(tool)
-                .map_err(|error| format!("invalid default tool core adapter ({tool}): {error}"))?;
-        }
-        if let Some(memory) = defaults.memory.as_deref() {
-            kernel
-                .set_default_core_memory_adapter(memory)
-                .map_err(|error| {
-                    format!("invalid default memory core adapter ({memory}): {error}")
-                })?;
-        }
-    }
-    Ok(())
-}
+#[cfg(test)]
+#[path = "spec_execution/bootstrap_policy_tests.rs"]
+mod bootstrap_policy_tests;
 
 #[cfg(test)]
-mod bootstrap_policy_tests {
-    use super::*;
-    use crate::spec_runtime::BootstrapSpec;
-
-    #[test]
-    fn bootstrap_policy_maps_distinct_acp_bridge_and_runtime_flags() {
-        let mut spec = RunnerSpec::template();
-        spec.bootstrap = Some(BootstrapSpec {
-            enabled: true,
-            allow_http_json_auto_apply: None,
-            allow_process_stdio_auto_apply: None,
-            allow_native_ffi_auto_apply: None,
-            allow_wasm_component_auto_apply: None,
-            allow_mcp_server_auto_apply: None,
-            allow_acp_bridge_auto_apply: Some(true),
-            allow_acp_runtime_auto_apply: Some(false),
-            block_unverified_high_risk_auto_apply: Some(true),
-            enforce_ready_execution: None,
-            max_tasks: None,
-        });
-
-        let policy = bootstrap_policy(&spec).expect("bootstrap policy should resolve");
-        assert!(policy.allow_acp_bridge_auto_apply);
-        assert!(!policy.allow_acp_runtime_auto_apply);
-        assert!(policy.block_unverified_high_risk_auto_apply);
-    }
-}
+#[path = "spec_execution/setup_readiness_context_tests.rs"]
+mod setup_readiness_context_tests;
 
 #[cfg(test)]
-mod setup_readiness_context_tests {
-    use super::*;
-    use crate::PluginSetupReadinessSpec;
-
-    #[test]
-    fn collect_verified_env_var_names_ignores_blank_names_and_values() {
-        let env_vars = vec![
-            (OsString::from("TAVILY_API_KEY"), OsString::from("secret")),
-            (OsString::from("EMPTY_VALUE"), OsString::from("   ")),
-            (OsString::from("   "), OsString::from("ignored")),
-        ];
-
-        let verified_env_vars = collect_verified_env_var_names(env_vars);
-
-        assert_eq!(
-            verified_env_vars,
-            BTreeSet::from(["TAVILY_API_KEY".to_owned()])
-        );
-    }
-
-    #[test]
-    fn collect_verified_env_var_names_preserves_non_blank_name_spelling() {
-        let env_vars = vec![
-            (OsString::from(" TAVILY_API_KEY"), OsString::from("secret")),
-            (OsString::from("TAVILY_API_KEY "), OsString::from("secret")),
-        ];
-
-        let verified_env_vars = collect_verified_env_var_names(env_vars);
-
-        assert_eq!(
-            verified_env_vars,
-            BTreeSet::from([" TAVILY_API_KEY".to_owned(), "TAVILY_API_KEY ".to_owned(),])
-        );
-    }
-
-    #[test]
-    fn resolve_plugin_setup_readiness_context_falls_back_to_process_env_when_unspecified() {
-        let env_vars = vec![
-            (OsString::from("TAVILY_API_KEY"), OsString::from("secret")),
-            (OsString::from("EMPTY_VALUE"), OsString::from("   ")),
-        ];
-
-        let context = resolve_plugin_setup_readiness_context(None, env_vars);
-
-        assert_eq!(
-            context.verified_env_vars,
-            BTreeSet::from(["TAVILY_API_KEY".to_owned()])
-        );
-        assert!(context.verified_config_keys.is_empty());
-    }
-
-    #[test]
-    fn resolve_plugin_setup_readiness_context_uses_explicit_values_without_env_inheritance() {
-        let readiness_spec = PluginSetupReadinessSpec {
-            inherit_process_env: false,
-            verified_env_vars: vec![" TAVILY_API_KEY ".to_owned(), "".to_owned()],
-            verified_config_keys: vec![
-                " tools.web_search.default_provider ".to_owned(),
-                "   ".to_owned(),
-            ],
-        };
-        let env_vars = vec![(OsString::from("SHOULD_NOT_BE_USED"), OsString::from("set"))];
-
-        let context = resolve_plugin_setup_readiness_context(Some(&readiness_spec), env_vars);
-
-        assert_eq!(
-            context.verified_env_vars,
-            BTreeSet::from(["TAVILY_API_KEY".to_owned()])
-        );
-        assert_eq!(
-            context.verified_config_keys,
-            BTreeSet::from(["tools.web_search.default_provider".to_owned()])
-        );
-    }
-
-    #[test]
-    fn resolve_plugin_setup_readiness_context_merges_process_env_when_requested() {
-        let readiness_spec = PluginSetupReadinessSpec {
-            inherit_process_env: true,
-            verified_env_vars: vec!["TAVILY_API_KEY".to_owned()],
-            verified_config_keys: vec!["tools.web_search.default_provider".to_owned()],
-        };
-        let env_vars = vec![(OsString::from("OPENAI_API_KEY"), OsString::from("set"))];
-
-        let context = resolve_plugin_setup_readiness_context(Some(&readiness_spec), env_vars);
-
-        assert_eq!(
-            context.verified_env_vars,
-            BTreeSet::from(["OPENAI_API_KEY".to_owned(), "TAVILY_API_KEY".to_owned(),])
-        );
-        assert_eq!(
-            context.verified_config_keys,
-            BTreeSet::from(["tools.web_search.default_provider".to_owned()])
-        );
-    }
-}
+#[path = "spec_execution/bridge_runtime_policy_tests.rs"]
+mod bridge_runtime_policy_tests;
 
 #[cfg(test)]
-mod bridge_runtime_policy_tests {
-    use super::*;
-
-    #[test]
-    fn bridge_runtime_policy_honors_raw_circuit_breaker_override_when_scan_is_disabled() {
-        let mut spec = RunnerSpec::template();
-        let (mut bridge, _source) =
-            load_bundled_bridge_support_policy("openclaw-ecosystem-balanced")
-                .expect("bundled bridge support should resolve");
-        let mut security_scan = SecurityScanSpec {
-            enabled: false,
-            ..SecurityScanSpec::default()
-        };
-        security_scan.runtime.bridge_circuit_breaker.enabled = false;
-        bridge.security_scan = Some(security_scan);
-        spec.bridge_support = Some(bridge);
-
-        let policy =
-            bridge_runtime_policy(&spec, None).expect("bridge runtime policy should build");
-
-        assert!(!policy.bridge_circuit_breaker.enabled);
-    }
-
-    #[test]
-    fn bridge_runtime_policy_rejects_invalid_circuit_breaker_override() {
-        let mut spec = RunnerSpec::template();
-        let (mut bridge, _source) =
-            load_bundled_bridge_support_policy("openclaw-ecosystem-balanced")
-                .expect("bundled bridge support should resolve");
-        let mut security_scan = SecurityScanSpec {
-            enabled: false,
-            ..SecurityScanSpec::default()
-        };
-        security_scan
-            .runtime
-            .bridge_circuit_breaker
-            .failure_threshold = 0;
-        bridge.security_scan = Some(security_scan);
-        spec.bridge_support = Some(bridge);
-
-        let error = bridge_runtime_policy(&spec, None)
-            .expect_err("invalid bridge circuit breaker policy should fail");
-
-        assert!(error.contains("failure_threshold"));
-    }
-
-    #[test]
-    fn bridge_runtime_policy_normalizes_guest_readable_config_keys() {
-        let mut spec = RunnerSpec::template();
-        let (mut bridge, _source) =
-            load_bundled_bridge_support_policy("openclaw-ecosystem-balanced")
-                .expect("bundled bridge support should resolve");
-        let mut security_scan = SecurityScanSpec {
-            enabled: true,
-            ..SecurityScanSpec::default()
-        };
-        security_scan.runtime.guest_readable_config_keys = vec![
-            " provider.region ".to_owned(),
-            "channel.mode".to_owned(),
-            "provider.region".to_owned(),
-        ];
-        bridge.security_scan = Some(security_scan);
-        spec.bridge_support = Some(bridge);
-        let security_scan = spec
-            .bridge_support
-            .as_ref()
-            .and_then(|bridge| bridge.security_scan.as_ref());
-
-        let policy = bridge_runtime_policy(&spec, security_scan)
-            .expect("bridge runtime policy should build");
-
-        assert_eq!(
-            policy.wasm_guest_readable_config_keys,
-            BTreeSet::from(["channel.mode".to_owned(), "provider.region".to_owned(),])
-        );
-    }
-
-    #[test]
-    fn bridge_runtime_policy_rejects_invalid_guest_readable_config_key_namespace() {
-        let mut spec = RunnerSpec::template();
-        let (mut bridge, _source) =
-            load_bundled_bridge_support_policy("openclaw-ecosystem-balanced")
-                .expect("bundled bridge support should resolve");
-        let mut security_scan = SecurityScanSpec {
-            enabled: true,
-            ..SecurityScanSpec::default()
-        };
-        security_scan.runtime.guest_readable_config_keys = vec!["region".to_owned()];
-        bridge.security_scan = Some(security_scan);
-        spec.bridge_support = Some(bridge);
-        let security_scan = spec
-            .bridge_support
-            .as_ref()
-            .and_then(|bridge| bridge.security_scan.as_ref());
-
-        let error = bridge_runtime_policy(&spec, security_scan)
-            .expect_err("invalid guest-readable config key should fail");
-
-        assert!(error.contains("guest_readable_config_keys"));
-        assert!(error.contains("region"));
-        assert!(error.contains(WASM_GUEST_CONFIG_PROVIDER_PREFIX));
-        assert!(error.contains(WASM_GUEST_CONFIG_CHANNEL_PREFIX));
-    }
-
-    #[test]
-    fn bridge_runtime_policy_rejects_guest_readable_config_key_with_inner_whitespace() {
-        let mut spec = RunnerSpec::template();
-        let (mut bridge, _source) =
-            load_bundled_bridge_support_policy("openclaw-ecosystem-balanced")
-                .expect("bundled bridge support should resolve");
-        let mut security_scan = SecurityScanSpec {
-            enabled: true,
-            ..SecurityScanSpec::default()
-        };
-        security_scan.runtime.guest_readable_config_keys = vec!["provider.region name".to_owned()];
-        bridge.security_scan = Some(security_scan);
-        spec.bridge_support = Some(bridge);
-        let security_scan = spec
-            .bridge_support
-            .as_ref()
-            .and_then(|bridge| bridge.security_scan.as_ref());
-
-        let error = bridge_runtime_policy(&spec, security_scan)
-            .expect_err("guest-readable config key with inner whitespace should fail");
-
-        assert!(error.contains("provider.region name"));
-        assert!(error.contains("guest_readable_config_keys"));
-    }
-
-    #[tokio::test]
-    async fn execute_spec_rejects_invalid_guest_readable_config_keys_before_plugin_scan() {
-        let mut spec = RunnerSpec::template();
-        let (mut bridge, _source) =
-            load_bundled_bridge_support_policy("openclaw-ecosystem-balanced")
-                .expect("bundled bridge support should resolve");
-        let mut security_scan = SecurityScanSpec {
-            enabled: true,
-            ..SecurityScanSpec::default()
-        };
-        security_scan.runtime.guest_readable_config_keys = vec!["region".to_owned()];
-        bridge.security_scan = Some(security_scan);
-        spec.bridge_support = Some(bridge);
-        spec.plugin_scan = Some(PluginScanSpec {
-            enabled: true,
-            roots: vec!["/definitely/missing/loong-plugin-root".to_owned()],
-        });
-
-        let report = execute_spec(&spec, false).await;
-        let blocked_reason = report
-            .blocked_reason
-            .expect("invalid bridge runtime policy should block the spec");
-
-        assert_eq!(report.operation_kind, "blocked");
-        assert!(blocked_reason.contains("bridge runtime policy is invalid"));
-        assert!(blocked_reason.contains("guest_readable_config_keys"));
-        assert!(!blocked_reason.contains("plugin scan failed"));
-    }
-}
-
-#[cfg(test)]
-mod plugin_metadata_tests {
-    use super::*;
-    use kernel::{
-        Capability, PluginBridgeKind, PluginCompatibility, PluginCompatibilityMode,
-        PluginContractDialect, PluginDescriptor, PluginIR, PluginManifest, PluginRuntimeProfile,
-        PluginScanReport, PluginSetup, PluginSetupMode, PluginSlotClaim, PluginSlotMode,
-        PluginSourceKind, PluginTranslationReport, PluginTrustTier,
-    };
-    use std::collections::{BTreeMap, BTreeSet};
-
-    fn test_descriptor(source_kind: PluginSourceKind) -> PluginDescriptor {
-        let path = match source_kind {
-            PluginSourceKind::PackageManifest => "/tmp/pkg/loong.plugin.json".to_owned(),
-            PluginSourceKind::EmbeddedSource => "/tmp/pkg/plugin.py".to_owned(),
-        };
-        let package_manifest_path = match source_kind {
-            PluginSourceKind::PackageManifest => Some(path.clone()),
-            PluginSourceKind::EmbeddedSource => None,
-        };
-        let language = match source_kind {
-            PluginSourceKind::PackageManifest => "manifest".to_owned(),
-            PluginSourceKind::EmbeddedSource => "py".to_owned(),
-        };
-
-        PluginDescriptor {
-            path,
-            source_kind,
-            dialect: match source_kind {
-                PluginSourceKind::PackageManifest => PluginContractDialect::LoongPackageManifest,
-                PluginSourceKind::EmbeddedSource => PluginContractDialect::LoongEmbeddedSource,
-            },
-            dialect_version: Some("v1alpha1".to_owned()),
-            compatibility_mode: PluginCompatibilityMode::Native,
-            package_root: "/tmp/pkg".to_owned(),
-            package_manifest_path,
-            language,
-            manifest: PluginManifest {
-                api_version: Some("v1alpha1".to_owned()),
-                version: Some("0.3.0".to_owned()),
-                plugin_id: "search-plugin".to_owned(),
-                provider_id: "search-provider".to_owned(),
-                connector_name: "search-connector".to_owned(),
-                channel_id: Some("primary".to_owned()),
-                endpoint: Some("https://example.com/search".to_owned()),
-                capabilities: BTreeSet::from([Capability::InvokeConnector]),
-                trust_tier: PluginTrustTier::VerifiedCommunity,
-                metadata: BTreeMap::new(),
-                summary: Some("Search plugin".to_owned()),
-                tags: vec!["search".to_owned()],
-                input_examples: Vec::new(),
-                output_examples: Vec::new(),
-                defer_loading: false,
-                setup: Some(PluginSetup {
-                    mode: PluginSetupMode::MetadataOnly,
-                    surface: Some("web_search".to_owned()),
-                    required_env_vars: vec!["TAVILY_API_KEY".to_owned()],
-                    recommended_env_vars: vec!["TEAM_TAVILY_KEY".to_owned()],
-                    required_config_keys: vec!["tools.web_search.default_provider".to_owned()],
-                    default_env_var: Some("TAVILY_API_KEY".to_owned()),
-                    docs_urls: vec!["https://docs.example.com/tavily".to_owned()],
-                    remediation: Some("set a Tavily credential before enabling search".to_owned()),
-                }),
-                slot_claims: vec![PluginSlotClaim {
-                    slot: "provider:web_search".to_owned(),
-                    key: "tavily".to_owned(),
-                    mode: PluginSlotMode::Exclusive,
-                }],
-                compatibility: Some(PluginCompatibility {
-                    host_api: Some("loong-plugin/v1".to_owned()),
-                    host_version_req: Some(">=0.1.0-alpha.1".to_owned()),
-                }),
-            },
-        }
-    }
-
-    fn test_translation(descriptor: &PluginDescriptor) -> PluginTranslationReport {
-        PluginTranslationReport {
-            translated_plugins: 1,
-            bridge_distribution: BTreeMap::from([("http_json".to_owned(), 1)]),
-            entries: vec![PluginIR {
-                manifest_api_version: descriptor.manifest.api_version.clone(),
-                plugin_version: descriptor.manifest.version.clone(),
-                dialect: descriptor.dialect,
-                dialect_version: descriptor.dialect_version.clone(),
-                compatibility_mode: descriptor.compatibility_mode,
-                plugin_id: descriptor.manifest.plugin_id.clone(),
-                provider_id: descriptor.manifest.provider_id.clone(),
-                connector_name: descriptor.manifest.connector_name.clone(),
-                channel_id: descriptor.manifest.channel_id.clone(),
-                endpoint: descriptor.manifest.endpoint.clone(),
-                capabilities: descriptor.manifest.capabilities.clone(),
-                trust_tier: descriptor.manifest.trust_tier,
-                metadata: descriptor.manifest.metadata.clone(),
-                source_path: descriptor.path.clone(),
-                source_kind: descriptor.source_kind,
-                package_root: descriptor.package_root.clone(),
-                package_manifest_path: descriptor.package_manifest_path.clone(),
-                diagnostic_findings: Vec::new(),
-                setup: descriptor.manifest.setup.clone(),
-                channel_bridge: None,
-                slot_claims: descriptor.manifest.slot_claims.clone(),
-                compatibility: descriptor.manifest.compatibility.clone(),
-                runtime: PluginRuntimeProfile {
-                    source_language: descriptor.language.clone(),
-                    bridge_kind: PluginBridgeKind::HttpJson,
-                    adapter_family: "http-adapter".to_owned(),
-                    entrypoint_hint: "https://example.com/search".to_owned(),
-                },
-            }],
-        }
-    }
-
-    fn test_channel_bridge_descriptor(source_kind: PluginSourceKind) -> PluginDescriptor {
-        let mut descriptor = test_descriptor(source_kind);
-
-        descriptor.manifest.plugin_id = "weixin-clawbot-bridge".to_owned();
-        descriptor.manifest.provider_id = "weixin-bridge".to_owned();
-        descriptor.manifest.connector_name = "weixin-clawbot-http".to_owned();
-        descriptor.manifest.channel_id = Some("weixin".to_owned());
-        descriptor.manifest.endpoint = Some("http://127.0.0.1:8091/bridge".to_owned());
-        descriptor.manifest.metadata = BTreeMap::from([
-            (
-                "transport_family".to_owned(),
-                "wechat_clawbot_ilink_bridge".to_owned(),
-            ),
-            (
-                "target_contract".to_owned(),
-                "weixin:<account>:contact:<id> | weixin:<account>:room:<id>".to_owned(),
-            ),
-            ("account_scope".to_owned(), "multi_account".to_owned()),
-        ]);
-        descriptor.manifest.setup = Some(PluginSetup {
-            mode: PluginSetupMode::MetadataOnly,
-            surface: Some("channel".to_owned()),
-            required_env_vars: vec!["WEIXIN_BRIDGE_URL".to_owned()],
-            recommended_env_vars: vec!["WEIXIN_BRIDGE_ACCESS_TOKEN".to_owned()],
-            required_config_keys: vec![
-                "weixin.enabled".to_owned(),
-                "weixin.bridge_url".to_owned(),
-                "weixin.bridge_access_token".to_owned(),
-            ],
-            default_env_var: Some("WEIXIN_BRIDGE_URL".to_owned()),
-            docs_urls: vec!["https://docs.example.com/weixin-bridge".to_owned()],
-            remediation: Some("configure the sanctioned weixin bridge contract".to_owned()),
-        });
-
-        descriptor
-    }
-
-    fn test_channel_bridge_translation(descriptor: &PluginDescriptor) -> PluginTranslationReport {
-        PluginTranslationReport {
-            translated_plugins: 1,
-            bridge_distribution: BTreeMap::from([("http_json".to_owned(), 1)]),
-            entries: vec![PluginIR {
-                manifest_api_version: descriptor.manifest.api_version.clone(),
-                plugin_version: descriptor.manifest.version.clone(),
-                dialect: descriptor.dialect,
-                dialect_version: descriptor.dialect_version.clone(),
-                compatibility_mode: descriptor.compatibility_mode,
-                plugin_id: descriptor.manifest.plugin_id.clone(),
-                provider_id: descriptor.manifest.provider_id.clone(),
-                connector_name: descriptor.manifest.connector_name.clone(),
-                channel_id: descriptor.manifest.channel_id.clone(),
-                endpoint: descriptor.manifest.endpoint.clone(),
-                capabilities: descriptor.manifest.capabilities.clone(),
-                trust_tier: descriptor.manifest.trust_tier,
-                metadata: descriptor.manifest.metadata.clone(),
-                source_path: descriptor.path.clone(),
-                source_kind: descriptor.source_kind,
-                package_root: descriptor.package_root.clone(),
-                package_manifest_path: descriptor.package_manifest_path.clone(),
-                diagnostic_findings: Vec::new(),
-                setup: descriptor.manifest.setup.clone(),
-                channel_bridge: Some(kernel::PluginChannelBridgeContract {
-                    channel_id: Some("weixin".to_owned()),
-                    setup_surface: Some("channel".to_owned()),
-                    transport_family: Some("wechat_clawbot_ilink_bridge".to_owned()),
-                    target_contract: Some(
-                        "weixin:<account>:contact:<id> | weixin:<account>:room:<id>".to_owned(),
-                    ),
-                    account_scope: Some("multi_account".to_owned()),
-                    runtime_contract: Some("loong_channel_bridge_v1".to_owned()),
-                    runtime_operations: vec!["send_message".to_owned(), "receive_batch".to_owned()],
-                    runtime_metadata_issues: Vec::new(),
-                    readiness: kernel::PluginChannelBridgeReadiness {
-                        ready: true,
-                        missing_fields: Vec::new(),
-                    },
-                }),
-                slot_claims: descriptor.manifest.slot_claims.clone(),
-                compatibility: descriptor.manifest.compatibility.clone(),
-                runtime: PluginRuntimeProfile {
-                    source_language: descriptor.language.clone(),
-                    bridge_kind: PluginBridgeKind::HttpJson,
-                    adapter_family: "channel-bridge".to_owned(),
-                    entrypoint_hint: "http://127.0.0.1:8091/bridge".to_owned(),
-                },
-            }],
-        }
-    }
-
-    #[test]
-    fn enrich_scan_report_adds_package_manifest_provenance_and_setup_metadata() {
-        let descriptor = test_descriptor(PluginSourceKind::PackageManifest);
-        let report = PluginScanReport {
-            scanned_files: 1,
-            matched_plugins: 1,
-            diagnostic_findings: Vec::new(),
-            descriptors: vec![descriptor.clone()],
-        };
-        let translation = test_translation(&descriptor);
-
-        let enriched = enrich_scan_report_with_translation(&report, &translation, None);
-        let metadata = &enriched.descriptors[0].manifest.metadata;
-
-        assert_eq!(
-            metadata.get("plugin_source_kind").map(String::as_str),
-            Some("package_manifest")
-        );
-        assert_eq!(
-            metadata.get("plugin_trust_tier").map(String::as_str),
-            Some("verified-community")
-        );
-        assert_eq!(
-            metadata.get("plugin_package_root").map(String::as_str),
-            Some("/tmp/pkg")
-        );
-        assert_eq!(
-            metadata
-                .get("plugin_provenance_summary")
-                .map(String::as_str),
-            Some("package_manifest:/tmp/pkg/loong.plugin.json")
-        );
-        assert_eq!(
-            metadata
-                .get("plugin_package_manifest_path")
-                .map(String::as_str),
-            Some("/tmp/pkg/loong.plugin.json")
-        );
-        assert_eq!(
-            metadata.get("plugin_setup_mode").map(String::as_str),
-            Some("metadata_only")
-        );
-        assert_eq!(
-            metadata.get("plugin_setup_surface").map(String::as_str),
-            Some("web_search")
-        );
-        assert_eq!(
-            metadata
-                .get("plugin_setup_default_env_var")
-                .map(String::as_str),
-            Some("TAVILY_API_KEY")
-        );
-        assert_eq!(
-            metadata
-                .get("plugin_setup_required_env_vars_json")
-                .map(String::as_str),
-            Some("[\"TAVILY_API_KEY\"]")
-        );
-        assert_eq!(
-            metadata.get("plugin_slot_claims_json").map(String::as_str),
-            Some("[{\"slot\":\"provider:web_search\",\"key\":\"tavily\",\"mode\":\"exclusive\"}]")
-        );
-        assert_eq!(
-            metadata
-                .get("plugin_manifest_api_version")
-                .map(String::as_str),
-            Some("v1alpha1")
-        );
-        assert_eq!(
-            metadata.get("plugin_version").map(String::as_str),
-            Some("0.3.0")
-        );
-        assert_eq!(
-            metadata
-                .get("plugin_compatibility_host_api")
-                .map(String::as_str),
-            Some("loong-plugin/v1")
-        );
-        assert_eq!(
-            metadata
-                .get("plugin_compatibility_host_version_req")
-                .map(String::as_str),
-            Some(">=0.1.0-alpha.1")
-        );
-    }
-
-    #[test]
-    fn enrich_scan_report_omits_package_manifest_path_for_source_fallback() {
-        let descriptor = test_descriptor(PluginSourceKind::EmbeddedSource);
-        let report = PluginScanReport {
-            scanned_files: 1,
-            matched_plugins: 1,
-            diagnostic_findings: Vec::new(),
-            descriptors: vec![descriptor.clone()],
-        };
-        let translation = test_translation(&descriptor);
-
-        let enriched = enrich_scan_report_with_translation(&report, &translation, None);
-        let metadata = &enriched.descriptors[0].manifest.metadata;
-
-        assert_eq!(
-            metadata.get("plugin_source_kind").map(String::as_str),
-            Some("embedded_source")
-        );
-        assert_eq!(
-            metadata.get("plugin_trust_tier").map(String::as_str),
-            Some("verified-community")
-        );
-        assert_eq!(
-            metadata.get("plugin_package_root").map(String::as_str),
-            Some("/tmp/pkg")
-        );
-        assert_eq!(
-            metadata
-                .get("plugin_provenance_summary")
-                .map(String::as_str),
-            Some("embedded_source:/tmp/pkg/plugin.py")
-        );
-        assert_eq!(
-            metadata.get("plugin_setup_mode").map(String::as_str),
-            Some("metadata_only")
-        );
-        assert!(
-            !metadata.contains_key("plugin_package_manifest_path"),
-            "source fallback should not synthesize a package manifest path"
-        );
-    }
-
-    #[test]
-    fn enrich_scan_report_overwrites_forged_package_manifest_provenance_metadata() {
-        let mut descriptor = test_descriptor(PluginSourceKind::PackageManifest);
-
-        descriptor.manifest.metadata.insert(
-            "plugin_source_path".to_owned(),
-            "/forged/source-path".to_owned(),
-        );
-        descriptor.manifest.metadata.insert(
-            "plugin_source_kind".to_owned(),
-            "embedded_source".to_owned(),
-        );
-        descriptor.manifest.metadata.insert(
-            "plugin_package_root".to_owned(),
-            "/forged/package-root".to_owned(),
-        );
-        descriptor.manifest.metadata.insert(
-            "plugin_package_manifest_path".to_owned(),
-            "/forged/package-manifest".to_owned(),
-        );
-        descriptor.manifest.metadata.insert(
-            "plugin_provenance_summary".to_owned(),
-            "forged:summary".to_owned(),
-        );
-        descriptor
-            .manifest
-            .metadata
-            .insert("plugin_trust_tier".to_owned(), "unverified".to_owned());
-
-        let report = PluginScanReport {
-            scanned_files: 1,
-            matched_plugins: 1,
-            diagnostic_findings: Vec::new(),
-            descriptors: vec![descriptor.clone()],
-        };
-        let translation = test_translation(&descriptor);
-
-        let enriched = enrich_scan_report_with_translation(&report, &translation, None);
-        let metadata = &enriched.descriptors[0].manifest.metadata;
-
-        assert_eq!(
-            metadata.get("plugin_source_path").map(String::as_str),
-            Some("/tmp/pkg/loong.plugin.json")
-        );
-        assert_eq!(
-            metadata.get("plugin_source_kind").map(String::as_str),
-            Some("package_manifest")
-        );
-        assert_eq!(
-            metadata.get("plugin_package_root").map(String::as_str),
-            Some("/tmp/pkg")
-        );
-        assert_eq!(
-            metadata
-                .get("plugin_provenance_summary")
-                .map(String::as_str),
-            Some("package_manifest:/tmp/pkg/loong.plugin.json")
-        );
-        assert_eq!(
-            metadata.get("plugin_trust_tier").map(String::as_str),
-            Some("verified-community")
-        );
-        assert_eq!(
-            metadata
-                .get("plugin_package_manifest_path")
-                .map(String::as_str),
-            Some("/tmp/pkg/loong.plugin.json")
-        );
-    }
-
-    #[test]
-    fn enrich_scan_report_clears_forged_package_manifest_path_for_source_fallback() {
-        let mut descriptor = test_descriptor(PluginSourceKind::EmbeddedSource);
-
-        descriptor.manifest.metadata.insert(
-            "plugin_source_path".to_owned(),
-            "/forged/source-path".to_owned(),
-        );
-        descriptor.manifest.metadata.insert(
-            "plugin_source_kind".to_owned(),
-            "package_manifest".to_owned(),
-        );
-        descriptor.manifest.metadata.insert(
-            "plugin_package_root".to_owned(),
-            "/forged/package-root".to_owned(),
-        );
-        descriptor.manifest.metadata.insert(
-            "plugin_package_manifest_path".to_owned(),
-            "/forged/package-manifest".to_owned(),
-        );
-        descriptor.manifest.metadata.insert(
-            "plugin_provenance_summary".to_owned(),
-            "forged:summary".to_owned(),
-        );
-        descriptor
-            .manifest
-            .metadata
-            .insert("plugin_trust_tier".to_owned(), "official".to_owned());
-
-        let report = PluginScanReport {
-            scanned_files: 1,
-            matched_plugins: 1,
-            diagnostic_findings: Vec::new(),
-            descriptors: vec![descriptor.clone()],
-        };
-        let translation = test_translation(&descriptor);
-
-        let enriched = enrich_scan_report_with_translation(&report, &translation, None);
-        let metadata = &enriched.descriptors[0].manifest.metadata;
-
-        assert_eq!(
-            metadata.get("plugin_source_path").map(String::as_str),
-            Some("/tmp/pkg/plugin.py")
-        );
-        assert_eq!(
-            metadata.get("plugin_source_kind").map(String::as_str),
-            Some("embedded_source")
-        );
-        assert_eq!(
-            metadata.get("plugin_package_root").map(String::as_str),
-            Some("/tmp/pkg")
-        );
-        assert_eq!(
-            metadata
-                .get("plugin_provenance_summary")
-                .map(String::as_str),
-            Some("embedded_source:/tmp/pkg/plugin.py")
-        );
-        assert_eq!(
-            metadata.get("plugin_trust_tier").map(String::as_str),
-            Some("verified-community")
-        );
-        assert!(
-            !metadata.contains_key("plugin_package_manifest_path"),
-            "source fallback should remove forged package manifest paths"
-        );
-    }
-
-    #[test]
-    fn enrich_scan_report_overrides_conflicting_ad_hoc_setup_metadata() {
-        let mut descriptor = test_descriptor(PluginSourceKind::PackageManifest);
-        descriptor
-            .manifest
-            .metadata
-            .insert("plugin_setup_mode".to_owned(), "governed_entry".to_owned());
-        descriptor.manifest.metadata.insert(
-            "plugin_setup_surface".to_owned(),
-            "legacy_surface".to_owned(),
-        );
-        descriptor.manifest.metadata.insert(
-            "plugin_setup_required_env_vars_json".to_owned(),
-            "[\"LEGACY_KEY\"]".to_owned(),
-        );
-
-        let report = PluginScanReport {
-            scanned_files: 1,
-            matched_plugins: 1,
-            diagnostic_findings: Vec::new(),
-            descriptors: vec![descriptor.clone()],
-        };
-        let translation = test_translation(&descriptor);
-
-        let enriched = enrich_scan_report_with_translation(&report, &translation, None);
-        let metadata = &enriched.descriptors[0].manifest.metadata;
-
-        assert_eq!(
-            metadata.get("plugin_setup_mode").map(String::as_str),
-            Some("metadata_only")
-        );
-        assert_eq!(
-            metadata.get("plugin_setup_surface").map(String::as_str),
-            Some("web_search")
-        );
-        assert_eq!(
-            metadata
-                .get("plugin_setup_required_env_vars_json")
-                .map(String::as_str),
-            Some("[\"TAVILY_API_KEY\"]")
-        );
-    }
-
-    #[test]
-    fn enrich_scan_report_adds_channel_bridge_contract_metadata() {
-        let descriptor = test_channel_bridge_descriptor(PluginSourceKind::PackageManifest);
-        let report = PluginScanReport {
-            scanned_files: 1,
-            matched_plugins: 1,
-            diagnostic_findings: Vec::new(),
-            descriptors: vec![descriptor.clone()],
-        };
-        let translation = test_channel_bridge_translation(&descriptor);
-
-        let enriched = enrich_scan_report_with_translation(&report, &translation, None);
-        let metadata = &enriched.descriptors[0].manifest.metadata;
-
-        assert_eq!(
-            metadata.get("plugin_channel_id").map(String::as_str),
-            Some("weixin")
-        );
-        assert_eq!(
-            metadata
-                .get("plugin_channel_bridge_transport_family")
-                .map(String::as_str),
-            Some("wechat_clawbot_ilink_bridge")
-        );
-        assert_eq!(
-            metadata
-                .get("plugin_channel_bridge_target_contract")
-                .map(String::as_str),
-            Some("weixin:<account>:contact:<id> | weixin:<account>:room:<id>")
-        );
-        assert_eq!(
-            metadata
-                .get("plugin_channel_bridge_account_scope")
-                .map(String::as_str),
-            Some("multi_account")
-        );
-        assert_eq!(
-            metadata
-                .get("plugin_channel_bridge_ready")
-                .map(String::as_str),
-            Some("true")
-        );
-        assert_eq!(
-            metadata
-                .get("plugin_channel_bridge_missing_fields_json")
-                .map(String::as_str),
-            Some("[]")
-        );
-    }
-}
+#[path = "spec_execution/plugin_metadata_tests.rs"]
+mod plugin_metadata_tests;
