+++ crates/spec/src/spec_runtime.rs
@@ -41,20 +41,43 @@ use wasmtime::{
 
 #[cfg(any(test, feature = "test-hooks"))]
 use crate::WEBHOOK_TEST_RETRY_STATE;
-use crate::spec_execution::{normalize_path_for_policy, resolve_plugin_relative_path};
+use crate::{normalize_path_for_policy, resolve_plugin_relative_path};
 
+mod adapter_stubs;
 #[cfg(test)]
 mod bridge_kind_tests;
+mod bridge_support;
+mod connector_circuit;
 mod dynamic_catalog;
 mod embedded_pi_harness;
 mod http_json_bridge;
 mod plugin_contract_types;
 mod process_stdio_bridge;
 mod wasm_cache;
+mod wasm_execution_support;
 mod wasm_host_abi;
 mod wasm_runtime_policy;
 #[cfg(test)]
 mod wasm_runtime_tests;
+pub use adapter_stubs::{
+    AcpBridgeRuntimeExtension, ClawMigrationToolExtension, CoreToolRuntime, CrmCoreConnector,
+    CrmGrpcCoreConnector, FallbackCoreRuntime, KvCoreMemory, NativeCoreRuntime,
+    ShieldedConnectorExtension, SqlAnalyticsToolExtension, VectorIndexMemoryExtension,
+    WebhookConnector,
+};
+pub use bridge_support::{
+    caller_from_payload, caller_is_allowed, default_bridge_adapter_family,
+    default_bridge_entrypoint, detect_provider_bridge_kind, is_process_command_allowed,
+    parse_bridge_kind_label, parse_process_args, provider_allowed_callers,
+    resolve_wasm_component_artifact_path, resolve_wasm_export_name,
+};
+pub use connector_circuit::{
+    ConnectorCircuitAcquireError, ConnectorCircuitPhase, ConnectorCircuitRuntimeState,
+    ProgrammaticCircuitPhase, ProgrammaticCircuitRuntimeState,
+    acquire_connector_circuit_slot_for_state, connector_circuit_phase_label,
+    connector_circuit_remaining_cooldown_ms, record_connector_circuit_outcome_for_state,
+    validate_connector_circuit_breaker_policy,
+};
 pub use dynamic_catalog::{DynamicCatalogConnector, bridge_execution_payload};
 pub(crate) use dynamic_catalog::{
     default_runtime_adapter_family, normalize_runtime_source_language,
@@ -63,6 +86,12 @@ pub(crate) use dynamic_catalog::{
 pub use embedded_pi_harness::EmbeddedPiHarness;
 pub use http_json_bridge::execute_http_json_bridge;
 pub use plugin_contract_types::*;
+pub(crate) use plugin_contract_types::{
+    activation_runtime_contract_checksum_hex, build_plugin_runtime_health_result,
+    encode_plugin_runtime_health_result, invalid_plugin_runtime_health_result,
+    parse_plugin_activation_runtime_contract, plugin_activation_runtime_contract_json,
+    plugin_activation_runtime_contract_value,
+};
 pub use process_stdio_bridge::{
     ProcessStdioExchangeOutcome, execute_process_stdio_bridge, run_process_stdio_json_line_exchange,
 };
@@ -72,6 +101,13 @@ use wasm_cache::{
     CachedWasmModule, WasmArtifactFileIdentity, WasmModuleCacheLookup, build_wasm_module_cache_key,
     insert_cached_wasm_module, lookup_cached_wasm_module, modified_unix_nanos,
     wasm_artifact_file_identity, wasm_module_cache_capacity, wasm_module_cache_max_bytes,
+};
+use wasm_execution_support::{
+    WasmEntrypointSignature, WasmEpochDeadlineController, WasmRunEvidence, WasmRunOutcome,
+    WasmRunResult, WasmRuntimeExecutionContext, boxed_wasm_run_failure, compile_wasm_module,
+    read_wasm_artifact_bytes, wasm_artifact_sha256_hex, wasm_bridge_request_payload,
+    wasm_cache_lookup_disabled, wasm_runtime_execution_evidence, wasm_runtime_failure_reason,
+    wasm_snapshot_from_store,
 };
 pub(crate) use wasm_host_abi::{
     WASM_GUEST_CONFIG_CHANNEL_PREFIX, WASM_GUEST_CONFIG_PROVIDER_PREFIX,
@@ -464,198 +500,6 @@ pub struct ProgrammaticBatchExecutionSummary {
     pub scheduler_wait_cycles: usize,
     pub dispatch_order: Vec<String>,
     pub priority_dispatch_counts: BTreeMap<String, usize>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum ConnectorCircuitPhase {
-    Closed,
-    Open,
-    HalfOpen,
-}
-
-#[derive(Debug, Clone)]
-pub struct ConnectorCircuitRuntimeState {
-    pub phase: ConnectorCircuitPhase,
-    pub consecutive_failures: usize,
-    pub open_until: Option<TokioInstant>,
-    pub half_open_remaining_calls: usize,
-    pub half_open_successes: usize,
-}
-
-impl Default for ConnectorCircuitRuntimeState {
-    fn default() -> Self {
-        Self {
-            phase: ConnectorCircuitPhase::Closed,
-            consecutive_failures: 0,
-            open_until: None,
-            half_open_remaining_calls: 0,
-            half_open_successes: 0,
-        }
-    }
-}
-
-pub type ProgrammaticCircuitPhase = ConnectorCircuitPhase;
-pub type ProgrammaticCircuitRuntimeState = ConnectorCircuitRuntimeState;
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum ConnectorCircuitAcquireError {
-    Open { remaining_cooldown_ms: u64 },
-    HalfOpenReopened,
-}
-
-pub fn validate_connector_circuit_breaker_policy(
-    policy: &ConnectorCircuitBreakerPolicy,
-    context: &str,
-) -> Result<(), String> {
-    if !policy.enabled {
-        return Ok(());
-    }
-    if policy.failure_threshold == 0 {
-        return Err(format!(
-            "{context} failure_threshold must be greater than 0"
-        ));
-    }
-    if policy.cooldown_ms == 0 {
-        return Err(format!("{context} cooldown_ms must be greater than 0"));
-    }
-    if policy.half_open_max_calls == 0 {
-        return Err(format!(
-            "{context} half_open_max_calls must be greater than 0"
-        ));
-    }
-    if policy.success_threshold == 0 {
-        return Err(format!(
-            "{context} success_threshold must be greater than 0"
-        ));
-    }
-    if policy.success_threshold > policy.half_open_max_calls {
-        return Err(format!(
-            "{context} success_threshold must be <= half_open_max_calls"
-        ));
-    }
-    Ok(())
-}
-
-pub const fn connector_circuit_phase_label(phase: ConnectorCircuitPhase) -> &'static str {
-    match phase {
-        ConnectorCircuitPhase::Closed => "closed",
-        ConnectorCircuitPhase::Open => "open",
-        ConnectorCircuitPhase::HalfOpen => "half_open",
-    }
-}
-
-pub fn connector_circuit_remaining_cooldown_ms(
-    state: &ConnectorCircuitRuntimeState,
-    now: TokioInstant,
-) -> Option<u64> {
-    let open_until = state.open_until?;
-    if open_until <= now {
-        return Some(0);
-    }
-
-    let remaining_duration = open_until.duration_since(now);
-    let remaining_ms = remaining_duration.as_millis();
-    let clamped_ms = remaining_ms.min(u128::from(u64::MAX));
-
-    Some(clamped_ms as u64)
-}
-
-pub fn acquire_connector_circuit_slot_for_state(
-    policy: &ConnectorCircuitBreakerPolicy,
-    state: &mut ConnectorCircuitRuntimeState,
-    now: TokioInstant,
-) -> Result<&'static str, ConnectorCircuitAcquireError> {
-    if !policy.enabled {
-        return Ok("disabled");
-    }
-
-    if state.phase == ConnectorCircuitPhase::Open {
-        let remaining_cooldown_ms = connector_circuit_remaining_cooldown_ms(state, now);
-        if let Some(remaining_cooldown_ms) = remaining_cooldown_ms
-            && remaining_cooldown_ms > 0
-        {
-            return Err(ConnectorCircuitAcquireError::Open {
-                remaining_cooldown_ms,
-            });
-        }
-
-        state.phase = ConnectorCircuitPhase::HalfOpen;
-        state.open_until = None;
-        state.half_open_remaining_calls = policy.half_open_max_calls;
-        state.half_open_successes = 0;
-    }
-
-    if state.phase == ConnectorCircuitPhase::HalfOpen {
-        if state.half_open_remaining_calls == 0 {
-            let reopen_deadline = now + Duration::from_millis(policy.cooldown_ms);
-
-            state.phase = ConnectorCircuitPhase::Open;
-            state.open_until = Some(reopen_deadline);
-            return Err(ConnectorCircuitAcquireError::HalfOpenReopened);
-        }
-
-        state.half_open_remaining_calls = state.half_open_remaining_calls.saturating_sub(1);
-        return Ok("half_open");
-    }
-
-    Ok("closed")
-}
-
-pub fn record_connector_circuit_outcome_for_state(
-    policy: &ConnectorCircuitBreakerPolicy,
-    state: &mut ConnectorCircuitRuntimeState,
-    success: bool,
-    now: TokioInstant,
-) -> &'static str {
-    if !policy.enabled {
-        return "disabled";
-    }
-
-    match state.phase {
-        ConnectorCircuitPhase::Closed => {
-            if success {
-                state.consecutive_failures = 0;
-            } else {
-                state.consecutive_failures = state.consecutive_failures.saturating_add(1);
-                if state.consecutive_failures >= policy.failure_threshold {
-                    let reopen_deadline = now + Duration::from_millis(policy.cooldown_ms);
-
-                    state.phase = ConnectorCircuitPhase::Open;
-                    state.open_until = Some(reopen_deadline);
-                    state.half_open_remaining_calls = 0;
-                    state.half_open_successes = 0;
-                }
-            }
-        }
-        ConnectorCircuitPhase::HalfOpen => {
-            if success {
-                state.half_open_successes = state.half_open_successes.saturating_add(1);
-                if state.half_open_successes >= policy.success_threshold {
-                    state.phase = ConnectorCircuitPhase::Closed;
-                    state.consecutive_failures = 0;
-                    state.open_until = None;
-                    state.half_open_remaining_calls = 0;
-                    state.half_open_successes = 0;
-                } else if state.half_open_remaining_calls == 0 {
-                    let reopen_deadline = now + Duration::from_millis(policy.cooldown_ms);
-
-                    state.phase = ConnectorCircuitPhase::Open;
-                    state.open_until = Some(reopen_deadline);
-                    state.half_open_successes = 0;
-                }
-            } else {
-                let reopen_deadline = now + Duration::from_millis(policy.cooldown_ms);
-
-                state.phase = ConnectorCircuitPhase::Open;
-                state.open_until = Some(reopen_deadline);
-                state.half_open_remaining_calls = 0;
-                state.half_open_successes = 0;
-            }
-        }
-        ConnectorCircuitPhase::Open => {}
-    }
-
-    connector_circuit_phase_label(state.phase)
 }
 
 pub fn default_tool_search_limit() -> usize {
@@ -1066,177 +910,6 @@ pub struct PluginRuntimeHealthResult {
     pub issue: Option<String>,
 }
 
-fn invalid_plugin_runtime_health_result(reason: String) -> PluginRuntimeHealthResult {
-    PluginRuntimeHealthResult {
-        status: "unknown".to_owned(),
-        circuit_enabled: false,
-        circuit_phase: "unknown".to_owned(),
-        consecutive_failures: 0,
-        half_open_remaining_calls: 0,
-        half_open_successes: 0,
-        last_failure_reason: None,
-        issue: Some(reason),
-    }
-}
-
-fn plugin_runtime_health_status(
-    circuit_enabled: bool,
-    circuit_phase: &str,
-    consecutive_failures: usize,
-) -> String {
-    if !circuit_enabled || circuit_phase == "disabled" {
-        return "disabled".to_owned();
-    }
-
-    if circuit_phase == "open" {
-        return "quarantined".to_owned();
-    }
-
-    if circuit_phase == "half_open" {
-        return "degraded".to_owned();
-    }
-
-    if circuit_phase == "closed" && consecutive_failures > 0 {
-        return "degraded".to_owned();
-    }
-
-    if circuit_phase == "closed" {
-        return "healthy".to_owned();
-    }
-
-    "unknown".to_owned()
-}
-
-fn build_plugin_runtime_health_result(
-    policy: &ConnectorCircuitBreakerPolicy,
-    circuit_phase: String,
-    consecutive_failures: usize,
-    half_open_remaining_calls: usize,
-    half_open_successes: usize,
-    last_failure_reason: Option<String>,
-) -> PluginRuntimeHealthResult {
-    let circuit_enabled = policy.enabled;
-    let status = plugin_runtime_health_status(
-        circuit_enabled,
-        circuit_phase.as_str(),
-        consecutive_failures,
-    );
-
-    PluginRuntimeHealthResult {
-        status,
-        circuit_enabled,
-        circuit_phase,
-        consecutive_failures,
-        half_open_remaining_calls,
-        half_open_successes,
-        last_failure_reason,
-        issue: None,
-    }
-}
-
-fn encode_plugin_runtime_health_result(
-    health: &PluginRuntimeHealthResult,
-) -> Result<String, String> {
-    serde_json::to_string(health)
-        .map_err(|error| format!("serialize plugin runtime health failed: {error}"))
-}
-
-pub(crate) fn activation_runtime_contract_checksum_hex(bytes: &[u8]) -> String {
-    const OFFSET_BASIS: u64 = 0xcbf29ce484222325;
-    const PRIME: u64 = 0x100000001b3;
-
-    let mut hash = OFFSET_BASIS;
-    for byte in bytes {
-        hash ^= u64::from(*byte);
-        hash = hash.wrapping_mul(PRIME);
-    }
-    format!("{hash:016x}")
-}
-
-pub(crate) fn plugin_activation_runtime_contract_value(
-    contract: &PluginActivationRuntimeContract,
-) -> Value {
-    let mut value = Map::new();
-    let plugin_id = contract.plugin_id.clone();
-    let source_path = contract.source_path.clone();
-    let source_kind = contract.source_kind.as_str().to_owned();
-    let dialect = contract.dialect.as_str().to_owned();
-    let compatibility_mode = contract.compatibility_mode.as_str().to_owned();
-    let bridge_kind = contract.bridge_kind.as_str().to_owned();
-    let adapter_family = contract.adapter_family.clone();
-    let entrypoint_hint = contract.entrypoint_hint.clone();
-    let source_language = contract.source_language.clone();
-
-    value.insert("plugin_id".to_owned(), Value::String(plugin_id));
-    value.insert("source_path".to_owned(), Value::String(source_path));
-    value.insert("source_kind".to_owned(), Value::String(source_kind));
-    value.insert("dialect".to_owned(), Value::String(dialect));
-    value.insert(
-        "compatibility_mode".to_owned(),
-        Value::String(compatibility_mode),
-    );
-    value.insert("bridge_kind".to_owned(), Value::String(bridge_kind));
-    value.insert("adapter_family".to_owned(), Value::String(adapter_family));
-    value.insert("entrypoint_hint".to_owned(), Value::String(entrypoint_hint));
-    value.insert("source_language".to_owned(), Value::String(source_language));
-
-    if let Some(dialect_version) = &contract.dialect_version {
-        let dialect_version = dialect_version.clone();
-
-        value.insert("dialect_version".to_owned(), Value::String(dialect_version));
-    }
-    if let Some(compatibility_shim) = &contract.compatibility_shim {
-        let mut compatibility_shim_value = Map::new();
-        let shim_id = compatibility_shim.shim_id.clone();
-        let family = compatibility_shim.family.clone();
-
-        compatibility_shim_value.insert("shim_id".to_owned(), Value::String(shim_id));
-        compatibility_shim_value.insert("family".to_owned(), Value::String(family));
-        value.insert(
-            "compatibility_shim".to_owned(),
-            Value::Object(compatibility_shim_value),
-        );
-    }
-    if let Some(compatibility) = &contract.compatibility {
-        let mut compatibility_value = Map::new();
-
-        if let Some(host_api) = &compatibility.host_api {
-            let host_api = host_api.clone();
-
-            compatibility_value.insert("host_api".to_owned(), Value::String(host_api));
-        }
-        if let Some(host_version_req) = &compatibility.host_version_req {
-            let host_version_req = host_version_req.clone();
-
-            compatibility_value.insert(
-                "host_version_req".to_owned(),
-                Value::String(host_version_req),
-            );
-        }
-
-        value.insert(
-            "compatibility".to_owned(),
-            Value::Object(compatibility_value),
-        );
-    }
-
-    Value::Object(value)
-}
-
-pub(crate) fn plugin_activation_runtime_contract_json(
-    contract: &PluginActivationRuntimeContract,
-) -> Result<String, serde_json::Error> {
-    serde_json::to_string(&plugin_activation_runtime_contract_value(contract))
-}
-
-fn parse_plugin_activation_runtime_source_kind(raw: &str) -> Option<PluginSourceKind> {
-    match raw {
-        "package_manifest" => Some(PluginSourceKind::PackageManifest),
-        "embedded_source" => Some(PluginSourceKind::EmbeddedSource),
-        _ => None,
-    }
-}
-
 pub(crate) fn parse_plugin_activation_runtime_dialect(raw: &str) -> Option<PluginContractDialect> {
     match raw {
         "loong_package_manifest" | "loong_claw_package_manifest" => {
@@ -1276,96 +949,6 @@ pub(crate) fn parse_plugin_activation_runtime_bridge_kind(raw: &str) -> Option<P
         "unknown" => Some(PluginBridgeKind::Unknown),
         _ => None,
     }
-}
-
-fn optional_contract_string_field(
-    object: &serde_json::Map<String, Value>,
-    key: &str,
-) -> Result<Option<String>, String> {
-    match object.get(key) {
-        None | Some(Value::Null) => Ok(None),
-        Some(Value::String(value)) => Ok(Some(value.clone())),
-        Some(_) => Err(format!(
-            "plugin activation contract field `{key}` must be a string"
-        )),
-    }
-}
-
-fn required_contract_string_field(
-    object: &serde_json::Map<String, Value>,
-    key: &str,
-) -> Result<String, String> {
-    optional_contract_string_field(object, key)?
-        .ok_or_else(|| format!("plugin activation contract field `{key}` is required"))
-}
-
-pub(crate) fn parse_plugin_activation_runtime_contract(
-    raw: &str,
-) -> Result<PluginActivationRuntimeContract, String> {
-    let value: Value = serde_json::from_str(raw).map_err(|error| {
-        format!("plugin activation contract payload must be valid JSON: {error}")
-    })?;
-    let object = value
-        .as_object()
-        .ok_or_else(|| "plugin activation contract payload must be a JSON object".to_owned())?;
-
-    let compatibility_shim = match object.get("compatibility_shim") {
-        None | Some(Value::Null) => None,
-        Some(Value::Object(shim)) => Some(PluginCompatibilityShim {
-            shim_id: required_contract_string_field(shim, "shim_id")?,
-            family: required_contract_string_field(shim, "family")?,
-        }),
-        Some(_) => {
-            return Err(
-                "plugin activation contract field `compatibility_shim` must be an object"
-                    .to_owned(),
-            );
-        }
-    };
-
-    let compatibility = match object.get("compatibility") {
-        None | Some(Value::Null) => None,
-        Some(Value::Object(compatibility)) => Some(PluginCompatibility {
-            host_api: optional_contract_string_field(compatibility, "host_api")?,
-            host_version_req: optional_contract_string_field(compatibility, "host_version_req")?,
-        }),
-        Some(_) => {
-            return Err(
-                "plugin activation contract field `compatibility` must be an object".to_owned(),
-            );
-        }
-    };
-
-    let source_kind_raw = required_contract_string_field(object, "source_kind")?;
-    let dialect_raw = required_contract_string_field(object, "dialect")?;
-    let compatibility_mode_raw = required_contract_string_field(object, "compatibility_mode")?;
-    let bridge_kind_raw = required_contract_string_field(object, "bridge_kind")?;
-
-    Ok(PluginActivationRuntimeContract {
-        plugin_id: required_contract_string_field(object, "plugin_id")?,
-        source_path: required_contract_string_field(object, "source_path")?,
-        source_kind: parse_plugin_activation_runtime_source_kind(&source_kind_raw).ok_or_else(
-            || format!("plugin activation contract field `source_kind` has unsupported value `{source_kind_raw}`"),
-        )?,
-        dialect: parse_plugin_activation_runtime_dialect(&dialect_raw).ok_or_else(|| {
-            format!("plugin activation contract field `dialect` has unsupported value `{dialect_raw}`")
-        })?,
-        dialect_version: optional_contract_string_field(object, "dialect_version")?,
-        compatibility_mode: parse_plugin_activation_runtime_mode(&compatibility_mode_raw)
-            .ok_or_else(|| {
-                format!(
-                    "plugin activation contract field `compatibility_mode` has unsupported value `{compatibility_mode_raw}`"
-                )
-            })?,
-        compatibility_shim,
-        bridge_kind: parse_plugin_activation_runtime_bridge_kind(&bridge_kind_raw).ok_or_else(
-            || format!("plugin activation contract field `bridge_kind` has unsupported value `{bridge_kind_raw}`"),
-        )?,
-        adapter_family: required_contract_string_field(object, "adapter_family")?,
-        entrypoint_hint: required_contract_string_field(object, "entrypoint_hint")?,
-        source_language: required_contract_string_field(object, "source_language")?,
-        compatibility,
-    })
 }
 
 #[derive(Debug, Default)]
@@ -1743,72 +1326,6 @@ impl HotfixSpec {
     }
 }
 
-pub struct WebhookConnector;
-
-#[async_trait]
-impl CoreConnectorAdapter for WebhookConnector {
-    fn name(&self) -> &str {
-        "webhook"
-    }
-
-    async fn invoke_core(
-        &self,
-        command: ConnectorCommand,
-    ) -> Result<ConnectorOutcome, ConnectorError> {
-        #[cfg(any(test, feature = "test-hooks"))]
-        if let Some(test_config) = command
-            .payload
-            .as_object()
-            .and_then(|payload| payload.get("_loong_test"))
-            .and_then(Value::as_object)
-        {
-            let delay_ms = test_config
-                .get("delay_ms")
-                .and_then(Value::as_u64)
-                .unwrap_or(0);
-            if delay_ms > 0 {
-                sleep(Duration::from_millis(delay_ms)).await;
-            }
-            let request_id = test_config
-                .get("request_id")
-                .and_then(Value::as_str)
-                .unwrap_or_default()
-                .trim()
-                .to_owned();
-            let failures_before_success = test_config
-                .get("failures_before_success")
-                .and_then(Value::as_u64)
-                .unwrap_or(0) as usize;
-            if !request_id.is_empty() && failures_before_success > 0 {
-                let attempts_map =
-                    WEBHOOK_TEST_RETRY_STATE.get_or_init(|| Mutex::new(BTreeMap::new()));
-                let current_attempt = {
-                    let mut guard = attempts_map.lock().map_err(|_err| {
-                        ConnectorError::Execution("retry test state mutex poisoned".to_owned())
-                    })?;
-                    let entry = guard.entry(request_id.clone()).or_insert(0);
-                    *entry = entry.saturating_add(1);
-                    *entry
-                };
-                if current_attempt <= failures_before_success {
-                    return Err(ConnectorError::Execution(format!(
-                        "simulated transient failure for request_id={request_id}, attempt={current_attempt}, threshold={failures_before_success}"
-                    )));
-                }
-            }
-        }
-
-        Ok(ConnectorOutcome {
-            status: "ok".to_owned(),
-            payload: json!({
-                "connector": "webhook",
-                "operation": command.operation,
-                "payload": command.payload,
-            }),
-        })
-    }
-}
-
 include!("spec_bridge_protocol.inc.rs");
 
 fn with_execution_security_tier(
@@ -1911,291 +1428,6 @@ fn resolve_expected_wasm_sha256(
     }
 
     Ok(resolved_pin)
-}
-
-fn wasm_artifact_sha256_hex(bytes: &[u8]) -> String {
-    let digest = Sha256::digest(bytes);
-    let mut out = String::with_capacity(64);
-    for byte in digest {
-        let _ = std::fmt::Write::write_fmt(&mut out, format_args!("{byte:02x}"));
-    }
-    out
-}
-
-#[derive(Debug)]
-struct WasmArtifactBytes {
-    bytes: Vec<u8>,
-    modified_unix_ns: Option<u128>,
-    file_identity: Option<WasmArtifactFileIdentity>,
-}
-
-fn read_wasm_artifact_bytes(artifact_path: &Path) -> Result<WasmArtifactBytes, String> {
-    let mut artifact_file = fs::File::open(artifact_path)
-        .map_err(|error| format!("failed to open wasm artifact: {error}"))?;
-    let artifact_metadata = artifact_file
-        .metadata()
-        .map_err(|error| format!("failed to read wasm artifact metadata: {error}"))?;
-    if !artifact_metadata.file_type().is_file() {
-        return Err("wasm artifact path must reference a regular file".to_owned());
-    }
-
-    let expected_size = artifact_metadata.len().min(8 * 1024 * 1024_u64) as usize;
-    let mut bytes = Vec::with_capacity(expected_size);
-    artifact_file
-        .read_to_end(&mut bytes)
-        .map_err(|error| format!("failed to read wasm artifact: {error}"))?;
-
-    Ok(WasmArtifactBytes {
-        bytes,
-        modified_unix_ns: modified_unix_nanos(&artifact_metadata),
-        file_identity: wasm_artifact_file_identity(&artifact_metadata),
-    })
-}
-
-fn compile_wasm_module(
-    module_bytes: &[u8],
-    fuel_enabled: bool,
-    epoch_interruption_enabled: bool,
-    artifact_sha256: Option<String>,
-) -> Result<CachedWasmModule, String> {
-    let mut config = WasmtimeConfig::new();
-    // On macOS, default to `false` because Wasmtime's signal-based trap path
-    // relies on a global machports handler thread, which has shown intermittent
-    // aborts under highly parallel bridge tests.
-    config.signals_based_traps(wasm_signals_based_traps_enabled_from_env());
-    if fuel_enabled {
-        config.consume_fuel(true);
-    }
-    if epoch_interruption_enabled {
-        config.epoch_interruption(true);
-    }
-    let engine = WasmtimeEngine::new(&config)
-        .map_err(|error| format!("failed to initialize wasmtime engine: {error}"))?;
-    let module = WasmtimeModule::new(&engine, module_bytes)
-        .map_err(|error| format!("failed to compile wasm module: {error}"))?;
-    Ok(CachedWasmModule {
-        engine,
-        module,
-        artifact_sha256,
-    })
-}
-
-#[derive(Debug)]
-struct WasmEpochDeadlineController {
-    cancel_tx: Option<mpsc::Sender<()>>,
-    worker: Option<thread::JoinHandle<()>>,
-}
-
-impl WasmEpochDeadlineController {
-    fn start(engine: &WasmtimeEngine, timeout_ms: u64) -> Result<Self, String> {
-        let (cancel_tx, cancel_rx) = mpsc::channel::<()>();
-        let timeout = Duration::from_millis(timeout_ms);
-        let engine = engine.clone();
-        let thread_name = "loong-wasm-timeout".to_owned();
-        let worker = thread::Builder::new()
-            .name(thread_name)
-            .spawn(move || {
-                let wait_result = cancel_rx.recv_timeout(timeout);
-                let timed_out = matches!(wait_result, Err(mpsc::RecvTimeoutError::Timeout));
-                if timed_out {
-                    engine.increment_epoch();
-                }
-            })
-            .map_err(|error| format!("failed to start wasm timeout watchdog: {error}"))?;
-        Ok(Self {
-            cancel_tx: Some(cancel_tx),
-            worker: Some(worker),
-        })
-    }
-
-    fn disarm(&mut self) {
-        if let Some(cancel_tx) = self.cancel_tx.take() {
-            let _ = cancel_tx.send(());
-        }
-        if let Some(worker) = self.worker.take() {
-            let _ = worker.join();
-        }
-    }
-}
-
-impl Drop for WasmEpochDeadlineController {
-    fn drop(&mut self) {
-        self.disarm();
-    }
-}
-
-fn wasm_runtime_failure_reason(
-    error: &wasmtime::Error,
-    timeout_ms: Option<u64>,
-    context: &str,
-) -> (String, bool) {
-    let is_interrupt_trap = matches!(
-        error.downcast_ref::<WasmtimeTrap>(),
-        Some(trap) if *trap == WasmtimeTrap::Interrupt
-    );
-
-    let Some(timeout_ms) = timeout_ms else {
-        return (format!("{context}: {error}"), false);
-    };
-    if !is_interrupt_trap {
-        return (format!("{context}: {error}"), false);
-    }
-
-    let timeout_reason = format!("wasm execution timed out after {timeout_ms}ms");
-    (timeout_reason, true)
-}
-
-fn wasm_cache_lookup_disabled(
-    cache_capacity: usize,
-    cache_max_bytes: usize,
-) -> WasmModuleCacheLookup {
-    WasmModuleCacheLookup {
-        hit: false,
-        inserted: false,
-        evicted_entries: 0,
-        cache_len: 0,
-        cache_capacity,
-        cache_total_module_bytes: 0,
-        cache_max_bytes,
-    }
-}
-
-#[derive(Debug, Clone, Copy)]
-enum WasmEntrypointSignature {
-    I32,
-    Unit,
-}
-
-impl WasmEntrypointSignature {
-    const fn as_str(self) -> &'static str {
-        match self {
-            Self::I32 => "() -> i32",
-            Self::Unit => "() -> ()",
-        }
-    }
-}
-
-#[derive(Debug, Clone, Default)]
-struct WasmRunEvidence {
-    entrypoint_signature: Option<&'static str>,
-    guest_exit_code: Option<i32>,
-    host_abi: WasmHostAbiSnapshot,
-}
-
-#[derive(Debug, Clone)]
-struct WasmRuntimeExecutionContext {
-    artifact_path: String,
-    export_name: String,
-    operation: String,
-    payload: Value,
-    request: Value,
-    module_size_bytes: usize,
-    fuel_limit: Option<u64>,
-    max_output_bytes: Option<usize>,
-    timeout_ms: Option<u64>,
-    cache_enabled: bool,
-    cache_lookup: WasmModuleCacheLookup,
-    cache_miss: bool,
-    expected_sha256: Option<String>,
-    artifact_sha256: Option<String>,
-}
-
-#[derive(Debug)]
-struct WasmRunOutcome {
-    consumed_fuel: Option<u64>,
-    timeout_triggered: bool,
-    evidence: WasmRunEvidence,
-}
-
-#[derive(Debug)]
-struct WasmRunFailure {
-    reason: String,
-    timeout_triggered: bool,
-    consumed_fuel: Option<u64>,
-    evidence: WasmRunEvidence,
-}
-
-type WasmRunResult<T> = Result<T, Box<WasmRunFailure>>;
-
-fn boxed_wasm_run_failure(
-    reason: impl Into<String>,
-    timeout_triggered: bool,
-    consumed_fuel: Option<u64>,
-    evidence: WasmRunEvidence,
-) -> Box<WasmRunFailure> {
-    Box::new(WasmRunFailure {
-        reason: reason.into(),
-        timeout_triggered,
-        consumed_fuel,
-        evidence,
-    })
-}
-
-fn wasm_bridge_request_payload(
-    provider: &kernel::ProviderConfig,
-    channel: &kernel::ChannelConfig,
-    command: &ConnectorCommand,
-) -> Value {
-    json!({
-        "provider_id": provider.provider_id,
-        "channel_id": channel.channel_id,
-        "operation": command.operation,
-        "payload": command.payload,
-    })
-}
-
-fn wasm_runtime_execution_evidence(
-    context: &WasmRuntimeExecutionContext,
-    timeout_triggered: bool,
-    fuel_consumed: Option<u64>,
-    evidence: &WasmRunEvidence,
-) -> Value {
-    let expected_sha256 = context.expected_sha256.clone();
-    let artifact_sha256 = context.artifact_sha256.clone();
-    let integrity_check_required = expected_sha256.is_some();
-    let integrity_check_passed = expected_sha256.is_none() || artifact_sha256.is_some();
-
-    json!({
-        "executor": "wasmtime_module",
-        "artifact_path": context.artifact_path,
-        "export": context.export_name,
-        "operation": context.operation,
-        "payload": context.payload,
-        "request": context.request,
-        "module_size_bytes": context.module_size_bytes,
-        "fuel_limit": context.fuel_limit,
-        "max_output_bytes": evidence.host_abi.max_output_bytes.or(context.max_output_bytes),
-        "timeout_ms": context.timeout_ms,
-        "timeout_triggered": timeout_triggered,
-        "fuel_consumed": fuel_consumed,
-        "cache_enabled": context.cache_enabled,
-        "cache_hit": context.cache_lookup.hit,
-        "cache_miss": context.cache_miss,
-        "cache_evicted_entries": context.cache_lookup.evicted_entries,
-        "cache_entries": context.cache_lookup.cache_len,
-        "cache_capacity": context.cache_lookup.cache_capacity,
-        "cache_total_module_bytes": context.cache_lookup.cache_total_module_bytes,
-        "cache_max_bytes": context.cache_lookup.cache_max_bytes,
-        "cache_inserted": context.cache_lookup.inserted,
-        "expected_sha256": expected_sha256,
-        "artifact_sha256": artifact_sha256,
-        "integrity_check_required": integrity_check_required,
-        "integrity_check_passed": integrity_check_passed,
-        "host_abi_enabled": evidence.host_abi.host_abi_enabled,
-        "entrypoint_signature": evidence.entrypoint_signature,
-        "guest_exit_code": evidence.guest_exit_code,
-        "guest_logs": evidence.host_abi.guest_logs,
-        "guest_logs_truncated": evidence.host_abi.guest_logs_truncated,
-        "output_text": evidence.host_abi.output_text,
-        "output_json": evidence.host_abi.output_json,
-    })
-}
-
-fn wasm_snapshot_from_store(
-    store: &WasmtimeStore<WasmHostAbiStoreData>,
-    host_abi_enabled: bool,
-) -> WasmHostAbiSnapshot {
-    store.data().snapshot(host_abi_enabled)
 }
 
 #[allow(clippy::indexing_slicing)] // serde_json::Value string-keyed IndexMut is infallible
@@ -2980,534 +2212,5 @@ pub fn execute_wasm_component_bridge(
             );
             execution
         }
-    }
-}
-
-pub fn resolve_wasm_component_artifact_path(
-    provider: &kernel::ProviderConfig,
-    channel_endpoint: &str,
-) -> Result<PathBuf, String> {
-    let raw = provider
-        .metadata
-        .get("component_resolved_path")
-        .cloned()
-        .or_else(|| provider.metadata.get("component_path").cloned())
-        .or_else(|| provider.metadata.get("component").cloned())
-        .or_else(|| {
-            let endpoint = channel_endpoint.trim();
-            endpoint
-                .to_ascii_lowercase()
-                .ends_with(".wasm")
-                .then(|| endpoint.to_owned())
-        })
-        .ok_or_else(|| "wasm_component execution requires component artifact path".to_owned())?;
-
-    if raw.starts_with("http://") || raw.starts_with("https://") {
-        return Err(
-            "wasm_component execution requires a local artifact path, remote URL is not allowed"
-                .to_owned(),
-        );
-    }
-
-    let candidate = PathBuf::from(&raw);
-    let resolved = if candidate.is_absolute() {
-        candidate
-    } else if let Some(source_path) = provider.metadata.get("plugin_source_path") {
-        resolve_plugin_relative_path(source_path, &raw)
-    } else {
-        candidate
-    };
-
-    Ok(normalize_path_for_policy(&resolved))
-}
-
-pub fn resolve_wasm_export_name(provider: &kernel::ProviderConfig) -> String {
-    let raw = provider
-        .metadata
-        .get("entrypoint")
-        .or_else(|| provider.metadata.get("entrypoint_hint"))
-        .cloned()
-        .unwrap_or_else(|| "run".to_owned());
-    raw.split([':', '/'])
-        .rfind(|segment| !segment.trim().is_empty())
-        .unwrap_or("run")
-        .to_owned()
-}
-
-pub fn parse_process_args(provider: &kernel::ProviderConfig) -> Vec<String> {
-    if let Some(args_json) = provider.metadata.get("args_json")
-        && let Ok(args) = serde_json::from_str::<Vec<String>>(args_json)
-    {
-        return args;
-    }
-
-    provider
-        .metadata
-        .get("args")
-        .map(|value| value.split_whitespace().map(str::to_owned).collect())
-        .unwrap_or_default()
-}
-
-pub fn provider_allowed_callers(provider: &kernel::ProviderConfig) -> BTreeSet<String> {
-    let mut allowed = BTreeSet::new();
-
-    if let Some(raw_json) = provider.metadata.get("allowed_callers_json")
-        && let Ok(values) = serde_json::from_str::<Vec<String>>(raw_json)
-    {
-        for value in values {
-            let normalized = value.trim().to_ascii_lowercase();
-            if !normalized.is_empty() {
-                allowed.insert(normalized);
-            }
-        }
-    }
-
-    if let Some(raw_list) = provider.metadata.get("allowed_callers") {
-        for token in raw_list.split([',', ';', ' ']) {
-            let normalized = token.trim().to_ascii_lowercase();
-            if !normalized.is_empty() {
-                allowed.insert(normalized);
-            }
-        }
-    }
-
-    allowed
-}
-
-pub fn caller_from_payload(payload: &Value) -> Option<String> {
-    payload
-        .get("_loong")
-        .and_then(Value::as_object)
-        .and_then(|meta| meta.get("caller"))
-        .and_then(Value::as_str)
-        .map(|caller| caller.trim().to_ascii_lowercase())
-        .filter(|caller| !caller.is_empty())
-}
-
-pub fn caller_is_allowed(caller: Option<&str>, allowed: &BTreeSet<String>) -> bool {
-    if allowed.is_empty() {
-        return true;
-    }
-    if allowed.contains("*") {
-        return true;
-    }
-    caller
-        .map(|value| value.trim().to_ascii_lowercase())
-        .is_some_and(|value| allowed.contains(&value))
-}
-
-pub fn is_process_command_allowed(program: &str, allowed: &BTreeSet<String>) -> bool {
-    loong_bridge_runtime::is_process_command_allowed(program, allowed)
-}
-
-pub fn detect_provider_bridge_kind(
-    provider: &kernel::ProviderConfig,
-    endpoint: &str,
-) -> PluginBridgeKind {
-    if let Some(raw) = provider.metadata.get("bridge_kind")
-        && let Some(kind) = parse_bridge_kind_label(raw)
-    {
-        return kind;
-    }
-
-    if endpoint.starts_with("http://") || endpoint.starts_with("https://") {
-        return PluginBridgeKind::HttpJson;
-    }
-
-    PluginBridgeKind::Unknown
-}
-
-pub fn parse_bridge_kind_label(raw: &str) -> Option<PluginBridgeKind> {
-    match raw.trim().to_ascii_lowercase().as_str() {
-        "http_json" | "http" => Some(PluginBridgeKind::HttpJson),
-        "process_stdio" | "stdio" => Some(PluginBridgeKind::ProcessStdio),
-        "native_ffi" | "ffi" => Some(PluginBridgeKind::NativeFfi),
-        "wasm_component" | "wasm" => Some(PluginBridgeKind::WasmComponent),
-        "mcp_server" | "mcp" => Some(PluginBridgeKind::McpServer),
-        "acp_bridge" | "acp" => Some(PluginBridgeKind::AcpBridge),
-        "acp_runtime" | "acpx" => Some(PluginBridgeKind::AcpRuntime),
-        "unknown" => Some(PluginBridgeKind::Unknown),
-        _ => None,
-    }
-}
-
-pub fn default_bridge_adapter_family(bridge_kind: PluginBridgeKind) -> String {
-    match bridge_kind {
-        PluginBridgeKind::HttpJson => "http-adapter".to_owned(),
-        PluginBridgeKind::ProcessStdio => "stdio-adapter".to_owned(),
-        PluginBridgeKind::NativeFfi => "ffi-adapter".to_owned(),
-        PluginBridgeKind::WasmComponent => "wasm-component-adapter".to_owned(),
-        PluginBridgeKind::McpServer => "mcp-adapter".to_owned(),
-        PluginBridgeKind::AcpBridge => "acp-bridge-adapter".to_owned(),
-        PluginBridgeKind::AcpRuntime => "acp-runtime-adapter".to_owned(),
-        PluginBridgeKind::Unknown => "unknown-adapter".to_owned(),
-    }
-}
-
-pub fn default_bridge_entrypoint(bridge_kind: PluginBridgeKind, endpoint: &str) -> String {
-    match bridge_kind {
-        PluginBridgeKind::HttpJson => endpoint.to_owned(),
-        PluginBridgeKind::ProcessStdio => "stdin/stdout::invoke".to_owned(),
-        PluginBridgeKind::NativeFfi => "lib::invoke".to_owned(),
-        PluginBridgeKind::WasmComponent => "component::run".to_owned(),
-        PluginBridgeKind::McpServer => "mcp::stdio".to_owned(),
-        PluginBridgeKind::AcpBridge => "acp::bridge".to_owned(),
-        PluginBridgeKind::AcpRuntime => "acp::turn".to_owned(),
-        PluginBridgeKind::Unknown => "unknown::invoke".to_owned(),
-    }
-}
-
-pub struct CrmCoreConnector;
-
-#[async_trait]
-impl CoreConnectorAdapter for CrmCoreConnector {
-    fn name(&self) -> &str {
-        "http-core"
-    }
-
-    async fn invoke_core(
-        &self,
-        command: ConnectorCommand,
-    ) -> Result<ConnectorOutcome, ConnectorError> {
-        Ok(ConnectorOutcome {
-            status: "ok".to_owned(),
-            payload: json!({
-                "tier": "core",
-                "adapter": "http-core",
-                "connector": command.connector_name,
-                "operation": command.operation,
-                "payload": command.payload,
-            }),
-        })
-    }
-}
-
-pub struct CrmGrpcCoreConnector;
-
-#[async_trait]
-impl CoreConnectorAdapter for CrmGrpcCoreConnector {
-    fn name(&self) -> &str {
-        "grpc-core"
-    }
-
-    async fn invoke_core(
-        &self,
-        command: ConnectorCommand,
-    ) -> Result<ConnectorOutcome, ConnectorError> {
-        Ok(ConnectorOutcome {
-            status: "ok".to_owned(),
-            payload: json!({
-                "tier": "core",
-                "adapter": "grpc-core",
-                "connector": command.connector_name,
-                "operation": command.operation,
-                "payload": command.payload,
-            }),
-        })
-    }
-}
-
-pub struct ShieldedConnectorExtension;
-
-#[async_trait]
-impl kernel::ConnectorExtensionAdapter for ShieldedConnectorExtension {
-    fn name(&self) -> &str {
-        "shielded-bridge"
-    }
-
-    async fn invoke_extension(
-        &self,
-        command: ConnectorCommand,
-        core: &(dyn CoreConnectorAdapter + Sync),
-    ) -> Result<ConnectorOutcome, ConnectorError> {
-        let probe = core
-            .invoke_core(ConnectorCommand {
-                connector_name: command.connector_name.clone(),
-                operation: "probe".to_owned(),
-                required_capabilities: BTreeSet::new(),
-                payload: json!({}),
-            })
-            .await?;
-        Ok(ConnectorOutcome {
-            status: "ok".to_owned(),
-            payload: json!({
-                "tier": "extension",
-                "extension": "shielded-bridge",
-                "operation": command.operation,
-                "core_probe": probe.payload,
-                "payload": command.payload,
-            }),
-        })
-    }
-}
-
-pub struct NativeCoreRuntime;
-
-#[async_trait]
-impl CoreRuntimeAdapter for NativeCoreRuntime {
-    fn name(&self) -> &str {
-        "native-core"
-    }
-
-    async fn execute_core(
-        &self,
-        request: RuntimeCoreRequest,
-    ) -> Result<RuntimeCoreOutcome, kernel::RuntimePlaneError> {
-        Ok(RuntimeCoreOutcome {
-            status: "ok".to_owned(),
-            payload: json!({
-                "adapter": "native-core",
-                "action": request.action,
-                "payload": request.payload,
-            }),
-        })
-    }
-}
-
-pub struct FallbackCoreRuntime;
-
-#[async_trait]
-impl CoreRuntimeAdapter for FallbackCoreRuntime {
-    fn name(&self) -> &str {
-        "fallback-core"
-    }
-
-    async fn execute_core(
-        &self,
-        request: RuntimeCoreRequest,
-    ) -> Result<RuntimeCoreOutcome, kernel::RuntimePlaneError> {
-        Ok(RuntimeCoreOutcome {
-            status: "ok".to_owned(),
-            payload: json!({
-                "adapter": "fallback-core",
-                "action": request.action,
-                "payload": request.payload,
-            }),
-        })
-    }
-}
-
-pub struct AcpBridgeRuntimeExtension;
-
-#[async_trait]
-impl RuntimeExtensionAdapter for AcpBridgeRuntimeExtension {
-    fn name(&self) -> &str {
-        "acp-bridge"
-    }
-
-    async fn execute_extension(
-        &self,
-        request: RuntimeExtensionRequest,
-        core: &(dyn CoreRuntimeAdapter + Sync),
-    ) -> Result<RuntimeExtensionOutcome, kernel::RuntimePlaneError> {
-        let core_probe = core
-            .execute_core(RuntimeCoreRequest {
-                action: "probe".to_owned(),
-                payload: json!({}),
-            })
-            .await?;
-        Ok(RuntimeExtensionOutcome {
-            status: "ok".to_owned(),
-            payload: json!({
-                "extension": "acp-bridge",
-                "action": request.action,
-                "core_probe": core_probe.payload,
-                "payload": request.payload,
-            }),
-        })
-    }
-}
-
-// Local stubs — spec adapters don't execute real tools/memory
-fn stub_tool_core(request: ToolCoreRequest) -> Result<ToolCoreOutcome, String> {
-    Ok(ToolCoreOutcome {
-        status: "ok".to_string(),
-        payload: json!({ "adapter": "core-tools", "tool": request.tool_name }),
-    })
-}
-
-fn maybe_execute_native_tool(
-    request: &ToolCoreRequest,
-    native_tool_executor: Option<crate::NativeToolExecutor>,
-) -> Option<Result<ToolCoreOutcome, String>> {
-    if let Some(executor) = native_tool_executor
-        && let Some(result) = executor(request.clone())
-    {
-        return Some(result);
-    }
-    if crate::tool_name_requires_native_tool_executor(request.tool_name.as_str()) {
-        return Some(Err(format!(
-            "native tool executor required for tool `{}`",
-            request.tool_name
-        )));
-    }
-    None
-}
-
-fn stub_memory_core(request: MemoryCoreRequest) -> Result<MemoryCoreOutcome, String> {
-    Ok(MemoryCoreOutcome {
-        status: "ok".to_string(),
-        payload: json!({ "adapter": "kv-core", "operation": request.operation }),
-    })
-}
-
-#[derive(Clone, Copy, Default)]
-pub struct CoreToolRuntime {
-    native_tool_executor: Option<crate::NativeToolExecutor>,
-}
-
-impl CoreToolRuntime {
-    pub const fn new(native_tool_executor: Option<crate::NativeToolExecutor>) -> Self {
-        Self {
-            native_tool_executor,
-        }
-    }
-}
-
-#[async_trait]
-impl CoreToolAdapter for CoreToolRuntime {
-    fn name(&self) -> &str {
-        "core-tools"
-    }
-
-    async fn execute_core_tool(
-        &self,
-        request: ToolCoreRequest,
-    ) -> Result<ToolCoreOutcome, kernel::ToolPlaneError> {
-        if let Some(result) = maybe_execute_native_tool(&request, self.native_tool_executor) {
-            return result.map_err(kernel::ToolPlaneError::Execution);
-        }
-        stub_tool_core(request).map_err(kernel::ToolPlaneError::Execution)
-    }
-}
-
-pub struct SqlAnalyticsToolExtension;
-
-#[async_trait]
-impl ToolExtensionAdapter for SqlAnalyticsToolExtension {
-    fn name(&self) -> &str {
-        "sql-analytics"
-    }
-
-    async fn execute_tool_extension(
-        &self,
-        request: ToolExtensionRequest,
-        core: &(dyn CoreToolAdapter + Sync),
-    ) -> Result<ToolExtensionOutcome, kernel::ToolPlaneError> {
-        let core_probe = core
-            .execute_core_tool(ToolCoreRequest {
-                tool_name: "schema_probe".to_owned(),
-                payload: json!({}),
-            })
-            .await?;
-        Ok(ToolExtensionOutcome {
-            status: "ok".to_owned(),
-            payload: json!({
-                "extension": "sql-analytics",
-                "action": request.extension_action,
-                "core_probe": core_probe.payload,
-                "payload": request.payload,
-            }),
-        })
-    }
-}
-
-pub struct ClawMigrationToolExtension;
-
-#[async_trait]
-impl ToolExtensionAdapter for ClawMigrationToolExtension {
-    fn name(&self) -> &str {
-        "claw-migration"
-    }
-
-    async fn execute_tool_extension(
-        &self,
-        request: ToolExtensionRequest,
-        core: &(dyn CoreToolAdapter + Sync),
-    ) -> Result<ToolExtensionOutcome, kernel::ToolPlaneError> {
-        let mut payload = request.payload.clone();
-        if payload.get("mode").is_none()
-            && let Some(object) = payload.as_object_mut()
-        {
-            object.insert(
-                "mode".to_owned(),
-                Value::String(request.extension_action.clone()),
-            );
-        }
-
-        let core_outcome = core
-            .execute_core_tool(ToolCoreRequest {
-                tool_name: "config.import".to_owned(),
-                payload,
-            })
-            .await?;
-        let mut response = serde_json::Map::new();
-        response.insert(
-            "extension".to_owned(),
-            Value::String("claw-migration".to_owned()),
-        );
-        response.insert(
-            "action".to_owned(),
-            Value::String(request.extension_action.clone()),
-        );
-        response.insert("core_outcome".to_owned(), core_outcome.payload.clone());
-        if let Some(core_object) = core_outcome.payload.as_object() {
-            for (key, value) in core_object {
-                response.entry(key.clone()).or_insert_with(|| value.clone());
-            }
-        } else {
-            response.insert("result".to_owned(), core_outcome.payload);
-        }
-        Ok(ToolExtensionOutcome {
-            status: "ok".to_owned(),
-            payload: Value::Object(response),
-        })
-    }
-}
-
-pub struct KvCoreMemory;
-
-#[async_trait]
-impl CoreMemoryAdapter for KvCoreMemory {
-    fn name(&self) -> &str {
-        "kv-core"
-    }
-
-    async fn execute_core_memory(
-        &self,
-        request: MemoryCoreRequest,
-    ) -> Result<MemoryCoreOutcome, kernel::MemoryPlaneError> {
-        stub_memory_core(request).map_err(kernel::MemoryPlaneError::Execution)
-    }
-}
-
-pub struct VectorIndexMemoryExtension;
-
-#[async_trait]
-impl MemoryExtensionAdapter for VectorIndexMemoryExtension {
-    fn name(&self) -> &str {
-        "vector-index"
-    }
-
-    async fn execute_memory_extension(
-        &self,
-        request: MemoryExtensionRequest,
-        core: &(dyn CoreMemoryAdapter + Sync),
-    ) -> Result<MemoryExtensionOutcome, kernel::MemoryPlaneError> {
-        let core_probe = core
-            .execute_core_memory(MemoryCoreRequest {
-                operation: "probe".to_owned(),
-                payload: json!({}),
-            })
-            .await?;
-        Ok(MemoryExtensionOutcome {
-            status: "ok".to_owned(),
-            payload: json!({
-                "extension": "vector-index",
-                "operation": request.operation,
-                "core_probe": core_probe.payload,
-                "payload": request.payload,
-            }),
-        })
     }
 }
