+++ crates/spec/src/spec_runtime/bridge_support.rs
@@ -1,0 +1,175 @@
+use super::*;
+
+pub fn resolve_wasm_component_artifact_path(
+    provider: &kernel::ProviderConfig,
+    channel_endpoint: &str,
+) -> Result<PathBuf, String> {
+    let raw = provider
+        .metadata
+        .get("component_resolved_path")
+        .cloned()
+        .or_else(|| provider.metadata.get("component_path").cloned())
+        .or_else(|| provider.metadata.get("component").cloned())
+        .or_else(|| {
+            let endpoint = channel_endpoint.trim();
+            endpoint
+                .to_ascii_lowercase()
+                .ends_with(".wasm")
+                .then(|| endpoint.to_owned())
+        })
+        .ok_or_else(|| "wasm_component execution requires component artifact path".to_owned())?;
+
+    if raw.starts_with("http://") || raw.starts_with("https://") {
+        return Err(
+            "wasm_component execution requires a local artifact path, remote URL is not allowed"
+                .to_owned(),
+        );
+    }
+
+    let candidate = PathBuf::from(&raw);
+    let resolved = if candidate.is_absolute() {
+        candidate
+    } else if let Some(source_path) = provider.metadata.get("plugin_source_path") {
+        resolve_plugin_relative_path(source_path, &raw)
+    } else {
+        candidate
+    };
+
+    Ok(normalize_path_for_policy(&resolved))
+}
+
+pub fn resolve_wasm_export_name(provider: &kernel::ProviderConfig) -> String {
+    let raw = provider
+        .metadata
+        .get("entrypoint")
+        .or_else(|| provider.metadata.get("entrypoint_hint"))
+        .cloned()
+        .unwrap_or_else(|| "run".to_owned());
+    raw.split([':', '/'])
+        .rfind(|segment| !segment.trim().is_empty())
+        .unwrap_or("run")
+        .to_owned()
+}
+
+pub fn parse_process_args(provider: &kernel::ProviderConfig) -> Vec<String> {
+    if let Some(args_json) = provider.metadata.get("args_json")
+        && let Ok(args) = serde_json::from_str::<Vec<String>>(args_json)
+    {
+        return args;
+    }
+
+    provider
+        .metadata
+        .get("args")
+        .map(|value| value.split_whitespace().map(str::to_owned).collect())
+        .unwrap_or_default()
+}
+
+pub fn provider_allowed_callers(provider: &kernel::ProviderConfig) -> BTreeSet<String> {
+    let mut allowed = BTreeSet::new();
+
+    if let Some(raw_json) = provider.metadata.get("allowed_callers_json")
+        && let Ok(values) = serde_json::from_str::<Vec<String>>(raw_json)
+    {
+        for value in values {
+            let normalized = value.trim().to_ascii_lowercase();
+            if !normalized.is_empty() {
+                allowed.insert(normalized);
+            }
+        }
+    }
+
+    if let Some(raw_list) = provider.metadata.get("allowed_callers") {
+        for token in raw_list.split([',', ';', ' ']) {
+            let normalized = token.trim().to_ascii_lowercase();
+            if !normalized.is_empty() {
+                allowed.insert(normalized);
+            }
+        }
+    }
+
+    allowed
+}
+
+pub fn caller_from_payload(payload: &Value) -> Option<String> {
+    payload
+        .get("_loong")
+        .and_then(Value::as_object)
+        .and_then(|meta| meta.get("caller"))
+        .and_then(Value::as_str)
+        .map(|caller| caller.trim().to_ascii_lowercase())
+        .filter(|caller| !caller.is_empty())
+}
+
+pub fn caller_is_allowed(caller: Option<&str>, allowed: &BTreeSet<String>) -> bool {
+    if allowed.is_empty() {
+        return true;
+    }
+    if allowed.contains("*") {
+        return true;
+    }
+    caller
+        .map(|value| value.trim().to_ascii_lowercase())
+        .is_some_and(|value| allowed.contains(&value))
+}
+
+pub fn is_process_command_allowed(program: &str, allowed: &BTreeSet<String>) -> bool {
+    loong_bridge_runtime::is_process_command_allowed(program, allowed)
+}
+
+pub fn detect_provider_bridge_kind(
+    provider: &kernel::ProviderConfig,
+    endpoint: &str,
+) -> PluginBridgeKind {
+    if let Some(raw) = provider.metadata.get("bridge_kind")
+        && let Some(kind) = parse_bridge_kind_label(raw)
+    {
+        return kind;
+    }
+
+    if endpoint.starts_with("http://") || endpoint.starts_with("https://") {
+        return PluginBridgeKind::HttpJson;
+    }
+
+    PluginBridgeKind::Unknown
+}
+
+pub fn parse_bridge_kind_label(raw: &str) -> Option<PluginBridgeKind> {
+    match raw.trim().to_ascii_lowercase().as_str() {
+        "http_json" | "http" => Some(PluginBridgeKind::HttpJson),
+        "process_stdio" | "stdio" => Some(PluginBridgeKind::ProcessStdio),
+        "native_ffi" | "ffi" => Some(PluginBridgeKind::NativeFfi),
+        "wasm_component" | "wasm" => Some(PluginBridgeKind::WasmComponent),
+        "mcp_server" | "mcp" => Some(PluginBridgeKind::McpServer),
+        "acp_bridge" | "acp" => Some(PluginBridgeKind::AcpBridge),
+        "acp_runtime" | "acpx" => Some(PluginBridgeKind::AcpRuntime),
+        "unknown" => Some(PluginBridgeKind::Unknown),
+        _ => None,
+    }
+}
+
+pub fn default_bridge_adapter_family(bridge_kind: PluginBridgeKind) -> String {
+    match bridge_kind {
+        PluginBridgeKind::HttpJson => "http-adapter".to_owned(),
+        PluginBridgeKind::ProcessStdio => "stdio-adapter".to_owned(),
+        PluginBridgeKind::NativeFfi => "ffi-adapter".to_owned(),
+        PluginBridgeKind::WasmComponent => "wasm-component-adapter".to_owned(),
+        PluginBridgeKind::McpServer => "mcp-adapter".to_owned(),
+        PluginBridgeKind::AcpBridge => "acp-bridge-adapter".to_owned(),
+        PluginBridgeKind::AcpRuntime => "acp-runtime-adapter".to_owned(),
+        PluginBridgeKind::Unknown => "unknown-adapter".to_owned(),
+    }
+}
+
+pub fn default_bridge_entrypoint(bridge_kind: PluginBridgeKind, endpoint: &str) -> String {
+    match bridge_kind {
+        PluginBridgeKind::HttpJson => endpoint.to_owned(),
+        PluginBridgeKind::ProcessStdio => "stdin/stdout::invoke".to_owned(),
+        PluginBridgeKind::NativeFfi => "lib::invoke".to_owned(),
+        PluginBridgeKind::WasmComponent => "component::run".to_owned(),
+        PluginBridgeKind::McpServer => "mcp::stdio".to_owned(),
+        PluginBridgeKind::AcpBridge => "acp::bridge".to_owned(),
+        PluginBridgeKind::AcpRuntime => "acp::turn".to_owned(),
+        PluginBridgeKind::Unknown => "unknown::invoke".to_owned(),
+    }
+}
